[roll] Roll fuchsia [rust_crates] Add hmac for usage by password_authenticator
We wish to use HMAC with a known value as a simple one-way function to
produce a low-entropy secret of a known length from the user's password
to use with credential manager in the pinweaver flow.  This helps
minimize the places distance the user's actual password gets transmitted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662936
Original-Revision: 33e4918dbf04ca20e9a62775a0cb23edeb723b49
GitOrigin-RevId: 3f516365f7219f154c748560ac08652f1abd28e9
Change-Id: I69e749887186bc1c803b911d450c820c821a72f6