[roll] Roll fuchsia [fuchsia.identity.credential] Adjust FIDL as appropriate
`bytes` is already a vector type, so vector<bytes>:1024 becomes a
vector-of-vector-of-byte, which is not what we want.

`reset_secret` is not something we have current plans to expose from
credential_manager at this time; since this is a FIDL table, it's enough
to just note to the reader that the value is not required to be
populated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662933
Original-Revision: c55f2063f3ee2efe118dfb55d4c58b9760a76aba
GitOrigin-RevId: 83fa329e764a03642243409eb448e0ff5d086955
Change-Id: I4a4d5dab9b2ad3a9b63c3924b0ea3501ab307c6d