[roll] Roll fuchsia [phys] Refactor phys stdout setup
This separates the code a bit so the uart library is not a direct
dependency of all phys setup.  Instead, it adds generic multiplexing
across serial and graphics.  This is used with the uart library in
existing code, but now in a dependency-injection style.

Original-Bug: 96043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662904
Original-Revision: 54dbc2a984d0e1ae43f796cbe465cced8180dd6c
GitOrigin-RevId: 6c97648021c001e1307cc185977d42b6c732749f
Change-Id: I39a728fa3ac3638f7f8346cfcc3b3efa093cfb0f