[roll] Roll fuchsia [audio] Add new thermal schema
The audio subsystem will be moving to a new version of the
API for thermal configuration. As part of this, we will be
changing how we represent our thermal configuration data,
in our static audio_core_config.json file. Essentially, the
"thermal_policy" section will be replaced by a new section
"thermal_states".

This CL adds the "thermal_states" section to the schema, so
that configs (some of which are in other repositories) can
be updated to include both old and new sections. Once this
is complete, AudioCore will change from referencing the old
section, to referencing the new. Once that is done, client
configs can be updated to remove the old "thermal_policy"
section completely.

Original-Bug: 94424

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662493
Original-Revision: 32c244998ae5c3506ac225f1d803215c23d3b001
GitOrigin-RevId: 28b8b92cbd5d2b0ec4abfa0a4159b4d6bde18907
Change-Id: I3f413a766d61921dfacd0b4cd306a8452d5f5ff4