[roll] Roll fuchsia [docs] Improve ffx docs syntax highlighting
"posix-terminal" for command line snippets that can be copied, removing the leading "$".
"none" for stuff that isn't a single format, such as a mix of a command line and free-text output.
"sh" for shell scripts - there are no shell scripts in this document, so removed all instances.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663222
Original-Revision: 8284e2f01420bb6781450e39b9a3d9869220584b
GitOrigin-RevId: 91ff8ade34a51fa61109c3df0648bca6981c9398
Change-Id: I5811d4372e5dac5468e68cfae93ea46cb0e6f9ff