[roll] Roll fuchsia [scrutiny] Remove memory leak created by Arc cycle EnginePlugin
The EnginePlugin holds a reference to the PluginManager that holds a
reference to the EnginePlugin when it's registered. This of course is
a cycle which creates a memory leak. Since Scrutiny holds the main
ownership of the Arc<Mutex<PluginManager>>, EnginePlugin now uses
Weak<Mutex<PluginManager>> so EnginePlugin only has a non-owning
reference.

Original-Bug: 88946
Original-Fixed: 96190
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661464
Original-Revision: e28d31b29ce1356214014e4f9d99a05c50c21207
GitOrigin-RevId: 4c712341a77c2efa436ea0816f99d2df3d05f694
Change-Id: Ia4df6f2f38133060145180a438d6e2036f3bbc29