[roll] Roll fuchsia [sdk-dart][realm-builder] child to child routing, events, and bug fixes
Connecting to services through ScopedInstance was incorrectly closing
the exposedDir, so it could not be used again. This also affected the
`connectToBinder()` call, if called after connecting to another service
via the `exposedDir`.

A new test that shows a realm with separate client and server components
requires starting the client eagerly. Since the test is not involved,
the realm has to be started, via `connectToBinder()`.

The only way to validate the test succeeds is to catch the component
`stopped` event (which should happen after the client starts, calls
`echoString()` on the server, and gets the server's response. So this
test shows how to connect to the framework `EventSource`, subscribe,
listen for the stopped event, and check for a successful status.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661895
Original-Revision: 088e8d77b1981720046aa77c1c265af9990afce6
GitOrigin-RevId: 6d6f9db1db7fb99e668b15bc8b7e0d8d9160d8f1
Change-Id: Iccc8c0cc21ef1c219a10fe68bee407723c6538c1