[roll] Roll fuchsia [netemul] Validate length of endpoint name
Because the name of a virtual endpoint in a netemul runner test is
reused for the interface name, in practice all endpoint names must be
<= fuchsia.net.interfaces/INTERFACE_NAME_LENGTH (15) characters long.

Currently, if a test configures an endpoint with a name longer than
that, the netemul client library will truncates the name on endpoint
creation, but the runner will still use the original longer name when
installing an interface in the netstack, causing the test to crash.
Instead, we should eagerly validate the endpoint name and only allow it
if it also works for the interface name.

Test: fx test -o netemul-runner-errors-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663207
Original-Revision: 2393903a1ca2793fcf14ae8906222c2a7efc9c6c
GitOrigin-RevId: 9834bf0d2566c01d9e747448f06014d3a603be85
Change-Id: I77e5d176725f2219afcf6d3c380abe81bc913e04