[roll] Roll fuchsia [fidl] Enable wire format v2 for overnet
In I4f95d937b0d3b828d2a059c7c671157c757f7f4e, support was added for the
FIDL persistence header, which can specify the encoded wire format which
is then used by the decoder to determine how to decode.

This follows on that CL by enabling the FIDL v2 wire format. The v1 wire
format is deprecated.

This should not depend on the CL enabling the persistence header from
rolling everywhere and could have been part of that CL. It is separated
out out of an abundance of caution.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663204
Original-Revision: 3e3f7a602739431fbfe44a9414983892cb712e56
GitOrigin-RevId: 158f1e9af1b9bb3ee8dfe6e40aeeab46ca8b4940
Change-Id: I80eb1499ebbb9e8f2c3a3957d2e50726b85bcf0f