[roll] Roll fuchsia [pwrbtn-monitor] Prefactor for OOM monitor changes
Refactor code so that more happens on the async loop. This moves
the device watching code to an async task. Previously this code
ran before the async loop started. This code waits forever on
targets without a power button.

This prefactor prepares for expanding the job of pwrbtn-monitor
to also listen for kernel OOM events. In a future change we'll
run the button and OOM listening code concurrently on the same
loop.

Original-Bug: 91708
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660807
Original-Revision: 284ba89e8e1e0316eadd0381887c3e945fd474a7
GitOrigin-RevId: a97e6ea87ffdc485a4f3117e834b3e2c6cb8531d
Change-Id: Id015de5b52229fbeb785905046877c19c1e769ff