[roll] Roll fuchsia [wlan] Use security authenticators in SME clients.
This change uses the `Authentication` FIDL message with bespoke
negotiation in code that communicates directly with SME (rather than the
Policy layer). `SecurityContext` types have been introduces in the
testing service utilities and the `wlandev` tool, which are used to
negotiate authentication using bespoke heuristics that maintain current
behavior in SME. Note that this is a preliminary change and should not
alter the behavior of the testing infrastructure nor `wlandev`.

This change was factored out of a larger abandoned change: fxr/643789.

Testing: (Refactored) unit tests and manual testing.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660009
Original-Revision: 11fd64a92168d9fd0da337b73ddf96f5413fde82
GitOrigin-RevId: 908f979429eb0c13ca5d5c09caf0d89944065f84
Change-Id: I9bf1bd8bfe35a218308684365c39230fcd3a239c