[roll] Roll fuchsia [kernel][vm] EnumerateChildren return zx_status_t
Changes EnumerateChildren, and its locked/internal variants, to return
an explicit zx_status_t instead of a bool. The intention is to allow
for disambiguating failures, specifically the case where the aspace is
in the destroyed state.

The actual VmEnumerator interface is unchanged and still returns bool
as there are not yet meaningful errors to expose here to justify
updating all the implementors.

Original-Bug: 94078
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660523
Original-Revision: fca4b454c299059eee0c5e848bc3775e591751c4
GitOrigin-RevId: e4fe8d7a209485164a1e8dfe0dbd8735d0e9d89f
Change-Id: Id2e73d2bc5b969e2719f4c9bdec617deba9a0987