[roll] Roll fuchsia Ensure public_deps is used to depend on rustc_library when appropriate.
Ensure our build system will work with upstream GN.

A few changes that will allow us to use upstream GN in the Fuchsia build
(we are currently stuck with an older revision due to a breaking
upstream GN change [1], see bug for details).

This ensures that `public_deps` is used instead of `deps` in group() and
source_set() targets that depend on rust_library() instances. Without
this their dependent cannot import the corresponding crate anymore.

Note that for all of this to work properly, the CL at [2] and its parent
must be submitted to GN first. I have verified manually that this CL fixes
the Fuchsia build in this case.

Original-Bug: 95999

[1] https://gn-review.googlesource.com/c/gn/+/13060
[2] https://gn-review.googlesource.com/c/gn/+/13280

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660607
Original-Revision: f81c8083142b85b98fa78bfaf8be39ea55afcee0
GitOrigin-RevId: f0dd828763de91a3c40d63360286a0714efabd40
Change-Id: Icaed7d3f834b82225a13f2508aaebf13183ab202