[roll] Roll fuchsia [kernel][job] Unlocked JobEnumerator
Changes JobDispatcher::EnumerateChildren to, if not recursing, call the
JobEnumerator methods without the parent JobDispatcher::lock_ held. Not
holding the lock has the effect of allowing the ZX_INFO_JOB_* queries
to be able to user copy without holding a lock. The actual set of
children enumerated remain a consistent snapshot as the set is gathered
with the lock held.

With the change to EnumerateChildren the method is split into two
methods, one for recursing and for direct descendants, instead of using
a bool since both the implementation and locking behavior are
sufficiently different for recursion.

Original-Bug: 94078
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660525
Original-Revision: 9203a9f003ca5ed5519ccdc41e2d3d246ef0ad7e
GitOrigin-RevId: 79c1bd50e80977103631b118c34fb9f391fc8fd4
Change-Id: Iadd58909897f8f14e209110ae8d343d81c4de1a1