[roll] Roll fuchsia [rndis-function] Drop lock before unbind callback
Once the unbind callback is triggered, the driver framework may call
release which deletes the driver. If we still hold the lock while this
happens then there is a use after free when release is called. We don't
need to hold the lock anyway, the callback doesn't need to be protected
and we can safely remove the debug assert (which did need the lock.)

Original-Fixed: 95777

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662632
Original-Revision: cf1a8637d3ff652f8b2868ce0c4d68c96b419753
GitOrigin-RevId: a3f6c2c1f35ad60b54e7102fb25cb0fb69e338d8
Change-Id: Ie94fa2df62f529942f0f7204d53f78704c7d1f2d