[roll] Roll fuchsia [swd-e2e-tests] Fix workstation e2e reboot tests
Previously, we tried to add a dependency on the target-side sl4f, which
the ota and reboot tests needs to work. Unfortunately the way we added
this only depended on the host-side parts of the sl4f package, which is
empty. This worked previously since we indirectly picked up a dependency
on sl4f from another target, which has since been removed.

This patch then adds an explicit dependency on sl4f that's not
parameterized by what platform we're building on, which fixes the
dependency. This gets the reboot test to pass, but the OTA tests are
blocked by other issues.

Original-Bug: 96481
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663423
Original-Revision: f67cab7877a32bc5e0e298024d6c5b93d38540e4
GitOrigin-RevId: ed73d6dce0c1e5ae3c26eb888a0a1a8743be1588
Change-Id: Ia2b131785ec5db398d2246f19bd252f1faf2f07a