[roll] Roll fuchsia [carnelian] Add a way to queue a message for a facet
In a recent discussion of the implementation of a progress bar,
it became clear that the plumbing needed to send a message
from a closure to a facet made it hard to build facets that can
do some work independently of the hosting view controller.

This change is a small that that makes matters a bit better. So long
as the facet is given a cloned app sender and records its facet ID, it
can send itself messages.

Testing: manual with the spinning square example.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663118
Original-Revision: b57dc10bfbb4d364cda23196bb35e0b7736b5017
GitOrigin-RevId: e0589e98d1b8817d323f141cdacbf29d531a7cf6
Change-Id: I85f79518c025ac13956945724420eabdf3f3bcde