[roll] Roll fuchsia [ffx] Refactor target wait.
This refactors target wait so that multiple points of failure can be
captured and store in logs, as well as provide a simple way to verify
that the target is discovered, and RCS is available.

The motivation for this change is in integration testing.

Original-Bug: 95702

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663643
Original-Revision: 0ac8163bbe26ff80d13c8bb3d5c87eac68f3f0d9
GitOrigin-RevId: 6b5136c293123f3ec94e7402590ae9d67e08c2b3
Change-Id: I1ac02761f276f1276dc9a7998fa1f0353c765841