[roll] Roll fuchsia [network/tests] Rewrite forwarding test to use sandbox
The test currently uses the netemul runner. It doesn't use any features
of the runner that the sandbox doesn't fulfill, and with the sandbox it
requires less configuration, is easier to parameterize over endpoint
type and netstack version, and is more in line with the rest of the
netstack integration tests.

Original-Bug: 96132
Original-Fixed: 96458
Original-Fixed: 96459
Test: fx test netstack-forwarding-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662827
Original-Revision: 4ffff24a38d378d1d8ed58edb0337bd8121edb7c
GitOrigin-RevId: 5504c5dd14c8686b2d12a66505b3550d870d6521
Change-Id: I32e1d8e08f0d60e2f6c7a88446b93acf546f6b03