[roll] Roll fuchsia [fidl][gidl] Validate malformed inline markers
Bit 0 of the envelope flag field is used to determine whether or not the
envelope is inlined or not. The other bits are unused and previously
unchecked in most bindings. However, a fuzzer run caught the lack of
validation of the unused bits, but all FIDL bits should be validated.

Original-Fixed: 96094
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661042
Original-Revision: a2cdb32c26762792763110ce4a38972f8abaad74
GitOrigin-RevId: 1f1686b38e2f3c95be206d3154fcd0937fac924a
Change-Id: I71ff0f176a08c13674f19386af5cf179de367ffe