[roll] Roll fuchsia [i915] Fix UB related to 0-length VLA
In C and C++, a VLA must have a positive, non-zero length. When
`CallOnDisplaysChanged` is invoked with no displays added and only
displays removed, the two VLAs in this function are declared to have
zero length, which invokes undefined behavior. Importantly, this
is a valid invocation for this function.

Original-Bug: 96550
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663782
Original-Revision: e04d6ce655532f0983164e27ebe51f7b64404e8d
GitOrigin-RevId: da3de2463ab0f4ed89384629cf2963044acc5499
Change-Id: I9200c7996060b844e86deb66f9b40a23dfa2c652