[roll] Roll fuchsia [wlan] Create network device library components
Create NetworkDevice and NetworkPort classes to make it easier to
implement support for the fuchsia.hardware.network.device protocol for
an improved data plane in wlan drivers.

This consists of two classes that a driver writer can use to represent a
device and one or more ports associated with that device. These classes
intend to help the driver writer working with network devices and
streamline the work specifically for wlan drivers.

Test: Included unit tests, separate experimental tests
Original-Bug: 79622
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567710
Original-Revision: 4fc47dd22932ab941a8df0e81983ac8f98d968b3
GitOrigin-RevId: e128201aa2eff65319148f6948bfe11361405446
Change-Id: I2fc35948507e614746a02fbe69058fbac123cc3a