[roll] Roll infra/recipes [build] Move out compdb filtering logic
This logic is only used by the firebase_docs recipe, which is slated for
eventual deletion, so it need not be centralized in the build recipe
module.

`build_results.filtered_compdb()` had been called by the kythe.py recipe
but it appears that the recipe doesn't actually require any filtering,
so it's fine to pass it the entire compdb file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/663221
Original-Revision: 62a060d75931e240d9aff9bfbc6968678b551ffb
GitOrigin-RevId: 6f7c3456c2410bb5f4ae855e379769912464600c
Change-Id: Ibafea1e66b51ce23f22c596d6e4d387d6e6f2763