[roll] Roll fuchsia [kernel][vm] Lookup gives offset on original object
Changes Lookup, and related methods that callback to a
VmObject::LookupFunction, to normalize the offset to be in terms of the
original object.

This has no impact to any existing callers, as all existing callers are
agnostic of whether the offsets were shifted or not due to slices, but
this provides a more consistent interface.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662640
Original-Revision: 1e976b3f2df5ce3768c44cf93ff124a1e0f12942
GitOrigin-RevId: 35f5c3810c34cfdfdf154bf48a36d4061f2021ee
Change-Id: Ie54bf52c91c8293e17a7d568446b48064dca1b26