[roll] Roll fuchsia [system-update-committer] Completely disable in recovery images.
system-update-committer already has logic to skip the health check if
the current partition is a recovery partition. Unfortunately, if the
paver is unable to determine the current partition (e.g. because A/B/R
metadata is corrupted, or when using the installer image),
system-update-committer would immediately reboot the system.

This CL adds a "disable" config option which is false by default, and
includes it in recovery images.

Original-Bug: 92116
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652702
Original-Revision: 4b4c8cf62e3a57827cf8c3b3e826ccff01f4e6dd
GitOrigin-RevId: b8001af0a2b306cb5800a60c0887b9ac31271d5d
Change-Id: Ib4da85e1522ff3621e235fdfb27630fdf45cf0d4