[roll] Roll fuchsia [ffx] Clean up ffx target when emu fails to start
The attached bug indicates a workflow where we were adding manual
targets to the ffx target collection, but failing to clean them up if we
encountered a later error.

This CL adds the clean-up step to that workflow, so that we don't leave
the manual targets behind.

Original-Bug: fxbug.dev/95967
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663506
Original-Revision: 8649bdec0b22d5bef020839013004925429d83b1
GitOrigin-RevId: a0a51d8726375b951d469321a169d15b74a12caf
Change-Id: I43aa4455e63346ea6b55b5fe739651ac72b72aa5