[roll] Roll fuchsia [wlan] Add pop_if to priority_queue
For convenience a user might need to remove frames from a priority queue
based on some attributes of those frames. As an example if a network
port goes down a driver might want to pop all frames from a queue that
belong to that port.

Add a method pop_if to priority_queue that takes a predicate that
evaluates each frame in the queue in priority order. If the predicate
returns true for a frame that frame will be popped.

Test: Included unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610863
Original-Revision: db3f1656c01b0c622812c53dc69d85db9818798f
GitOrigin-RevId: 31526f77ff2c1f7a9cba891a057b04916e45ae1e
Change-Id: Ic96dd1a3c161af86837a8bb29fe1cbf7b165809e