[roll] Roll fuchsia [ffx] Require an explicit flag when reusing emu
When an emulator instance is terminated but not cleaned up, the ffx
launcher can reuse the old instance's image files. This can be very
useful for certain workflows, but can also lead to some unexpected
behavior if the user doesn't know it's happening.

This CL adds a "--reuse" flag to the launcher. If the user attempts to
start an emulator with the same name as a persisted instance, it will
overwrite the existing images unless the "--reuse" flag is included on
the command line. This ensures that any reuse of old images is
intentional and expected.

Original-Bug: fxbug.dev/94626
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663115
Original-Revision: 3311d8cefbc4c064d94edc12db9474f7106f7617
GitOrigin-RevId: f74d18fe2372052a00655f5cb55a5760d5fe4aad
Change-Id: If63a09a1f9cb975ea04b8869372e29a85b27a2ef