[roll] Roll fuchsia [sdk-dart][realm-builder] get/replaceComponentDecl, improve start/stop
This fixes an outdated test that now uses newly-enabled start event
handling to demonstrate late-binding to a child component (starting the
component after building the realm) and waiting for the component to start.

It also improves the event callback support to allow a test to subscribe
to more than one event (for example, both "started" and "stopped"), which
wasn't possible with the EventType-specific constructors.

Adds support for component decl manipulation, and demonstrates adding
the Binder protocol to a component that didn't expose it in CML.

Also adds support for get/replaceRealmDecl, and replaceConfigValue (just
via a FIDL ValueSpec for now). Tests for replaceConfigValue will need to
be added in a future CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663803
Original-Revision: 06336a7639d17b87840298b9b6b9c3cf8a59cfc9
GitOrigin-RevId: ba584108dfb085dfdae689fd08c342ed10c533bd
Change-Id: I1e661e2fb4281276126f2f21074b067e225babb2