[roll] Roll fuchsia [starnix][signals] Move SignalActions to ThreadGroup
On Posix, signal handler are associated to the process and not the
thread. This CL reflects this by moving the ownership of the signal
handler table to the ThreadGroup. There is no behaviour change, as
previously, the code was making sure to correctly share the signal
handler table.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663726
Original-Revision: 77f6b8aed1022dc79710146d1ac94a7bf6522ea7
GitOrigin-RevId: ddf0c0f9e703aa430ea642073cd1062152283851
Change-Id: Ie614454fba575f70b5ba122a8c2ace6fe91a4505