[roll] Roll fuchsia [omaha-client] Decorate outgoing omaha-client requests with CUPv2 parameters
This behavior is gated on the presence of public keys in the Omaha
library config, and on an outgoing request setting |cup_sign_requests|.
If both of these conditions are met, RequestBuilders will be built with
cup handlers which can decorate outgoing requests.

This CL does not set cup_sign_requests anywhere, nor provide a path to
set it at runtime. This CL does not verify incoming requests, not store
the request metadata necessary to do so.

As a part of this CL, StandardCupv2Handler now accepts/stores references
to public keys instead of copying them, since we plan to construct it
per-request.

Also as a part of this CL, build_intermediate() attaches the cupv2
request parameters (if configured) and so returns a Result<_>, so
references and callsites are changed accordingly.

Original-Bug: 95799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663746
Original-Revision: 7915048a32968166d1c39bf02df7e26fe4330a90
GitOrigin-RevId: 872f1ed2b0427e9537dfdbf66472768aff2ed308
Change-Id: If0123c8a694b61429174844257eb55f71cfc5d71