[roll] Roll fuchsia [debugger] Fix global variable lookup.
Fix handling of DW_OP_addrx and DW_OP_constx which are used for some
global variables. Previously the code interpreted the parameter as a
byte offset into the table, but it should actually be an index into the
compilation unit's table (indicated by the byte offset addr_base).

Also enhance the expression printing for these operations in the
"sym-info" command which I used for diagnosing this problem.

(For background, this is a DWARF 5 addition and the implementation was
written before we had a good example to test with. Now the compiler
emits this operation frequently so the issue comes up.)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663025
Original-Revision: 6c0972d9864999f4f8ee39a5cca7dcbdb115aec3
GitOrigin-RevId: d257b33bfaff4bfe0dbbd500f2ee1afb3a1656cf
Change-Id: I5ea05585b7fdd0fbe1361aebf3f23ebf15b45538