[roll] Roll fuchsia [network/tests] Bind to address before client connects
The test currently involves a potential race between the server binding
to the address and the client attempting to connect to that address; if
the client wins, it'll see a "connection refused" error. Instead, the
server should bind to the address before allowing the client to connect.

Original-Fixed: 96597
Multiply: netstack-forwarding-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664284
Original-Revision: 6022a1145b1870683bb282974f009bbfa6f69495
GitOrigin-RevId: 6a8c92d366751d728bfc82193645dcb8bd013516
Change-Id: I6a4714f7678c5292a1eb196faeb58394fa4bfae3