[roll] Roll fuchsia [ffx] Fix assumptions in the emu config flags
The attached bug indicates a workflow where the current QEMU flags
failed. This is due to unstated assumptions that were made for previous
versions of QEMU that no longer hold in the current version. The result
was that what we thought were defaults actually aren't, and so emulator
start-up failed in certain circumstances.

This CL adjusts the flags needed when running with --headless. This has
been tested on FEMU and QEMU5, over ssh (no display) and CRD (with and
without --headless). The result is that FEMU behavior is unchanged, but
QEMU5 no longer fails over ssh.

An additional side effect is that QEMU5 does not open a window with the
monitor console anymore, but a user can still access the monitor console
by passing the --monitor flag during start-up.

Original-Bug: fxbug.dev/95860
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663960
Original-Revision: d69279ad41c8ee0f53ce02ee616abe11d43fa203
GitOrigin-RevId: 9c7cfa57cc8d67966c6b0b09d838168c37793e35
Change-Id: I9fa961482beeff9f3b66601a23c775058904a661