[roll] Roll fuchsia [credential_manager] Implements Add, Remove & Check Credential.
This change implements the Add, Remove and Check Credential operations for the
CredentialManager. The CredentialManager acts as the storage for the cr50
holding the HashTree and the encrypted CredentialMetadata.

On launch the CredentialManager checks if a HashTree already has been
persisted to disk and if so loads it into memory. Otherwise it will create
and persist a fresh HashTree and call ResetTree on the CR50.

Once the HashTree has been provisioned the CredentialManager starts
serving the fuchsia.identity.credential.CredentialManager protocol with
the three new operations.

Operations:
- AddCredential: Allocates a new credential in the hash tree binding a
  low-entropy secret to a high entropy secret in the CR50.
- CheckCredential: Validates an existing credential, if valid returns
  the high entropy secret.
- RemoveCredential: Removes an existing credential from the system.

Testing:
The vast bulk of this CL is test code testing the interactions of the
credential manager with a mocked pinweaver protocol.

Original-Fixed: 92205
Original-Fixed: 89058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634604
Original-Revision: e860c69dbedef7944855342ad9c1c77ecdc35fe9
GitOrigin-RevId: c197572448a864268463148b7b6c617666c66add
Change-Id: Ie58dc9ad60904e0795c6464ba11b08a711ae9fca