[roll] Roll fuchsia [omaha-client] Pass request metadata to omaha state machine.
This CL modifies the return types of build() and build_intermediate() to
pass the CUPv2 request metadata up a level to the Omaha state machine,
which builds and sends requests.

A follow-on CL will use this metadata to validate the response.

Original-Bug: 95799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664285
Original-Revision: 1ceae61db55c02ec12de076480a7205b05220ac7
GitOrigin-RevId: 8b7d8d9326c07611876b745fa13613945bfb2ea8
Change-Id: Ie6a6d8440837299ae8f7bf5c9fd60d503e5303cf