[roll] Roll fuchsia [phys] Randomized kernel load address test
This makes the trampoline boot test, use the input seed to
pick a random valid location where the kernel can be loaded.

The process to pick a location is as follows:
- From all memory ranges remove those that the bootloader,
  and use consider restricted. (Reserved, Peripheral).
- From this ranges, coalesce them if they are contiguous.
- From the coalesced ranges, remove all where an aligned
  chunk for the kernel cannot be picked, and we call
  this the candidated ranges.

A candidate range has N(> 1) slots for the kernel to be loaded at.
Each slot consisted of an aligned address within the range. Slots
are enumerated from the lowest address to the highest:

range.addr    2nd slot
  |     |      |       .......  |
      first slot               range.addr + range.size

Then the randomization is performed as follow:
 - Pick a random candidate.
 - From that candidate pick a random slot.

At this point we have a valid chunk where we can load the kernel
for trampoline boot.

Just for readability the code is broken down in such pieces.

Also BootZbi::GetKernelAllocationSize has been added, such
that it does all the offset calculations without creating
a BootZbi to obtain the actual memory needed to load the kernel,
which include the reserve memory.

Unlike KernelMemorySize, this does not include the 32 MB chunk,
and does not require us to move the kernel item into a BootZbi.

Test: fx run-zbi-test trampoline-boot-zbi-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646730
Original-Revision: a4e9b2a3ab2f7b08476a4512914e6c2eb5bfc15b
GitOrigin-RevId: cdf183eed68f99f117109691c5909eaf1945f1a4
Change-Id: Ie45fd81064d8bf365c385382f3482838e9fd7c4b