[roll] Roll fuchsia [camera][gym] Add frame arrival deadline (re-land)
Original fxr/656783 as a basis.

For Thermal testing, it is important to fail the test run if there
are indications of catastrophic failures such as no frames arriving.

This CL adds a deadline for frame arrival such a new deadline is
established at the beginning of each stream and at each subsequent
frame. The tolerance for lateness is a constant 5 seconds for all
streams (regardless of the frame rate). Please note that the check
only occurs every 2 seconds, so the watchdog timeout could occur as
late as 7 seconds after the stream stopped.

When this deadline is tripped (for any stream), camera-gym will
assert and crash out.

The thermal test rig will detect this failure and fail accordingly.

Original-Bug: b/219806158
Test: Run thermal test normally.
      Run thermal test with forced sensor reset (test should fail).
      Run normal camera-gym with forced frame drop.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662923
Original-Revision: 3627944e1bbb336c29d4f98891f29a69d71faffc
GitOrigin-RevId: f2b60eed7b6e75f6bbf0370bbc5d0f1ec5cad730
Change-Id: I1e54556895f32de24c1ceaa6b92a3fc932662971