[roll] Roll fuchsia [starnix] Add galaxies
This creates 4 different galaxies in which the starnix runner
operates. Each galaxy contains its own configuration and system image.
Each galaxy is exposed as a runner. For example, the "android" galaxy is
exposed as the "stardroid" runner, and the "android test" galaxy is
exposed as the "stardroid_test_runner".

This also changes the component hieararchy of the test runners. More
details in the README.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662283
Original-Revision: 4ec419bef18c5e628f6067c19189bf4c90f8bd7f
GitOrigin-RevId: c035a650115d5daa96c736f396df270ac40f20f8
Change-Id: Ib4dbc5e2a43ff3bad30b5cf264bccb5b7e4b31bd