[roll] Roll fuchsia [omaha-client] Soften opt-out logspam
fuchsia.update.config.OptOut is a protocol provided by some products and
not others. When connecting to a protocol that has no provider,
component manager currently logs a warning indicating it could not route
the connection to a provider, which looks scary on products that simply
do not have this component.

This change adds more logging to omaha-client when trying to use this
protocol to indicate that the protocol is not expected to exist on all
products in an effort to avoid that warning being a red herring in
investigations.

Eventually, CF will allow protocols to be marked as optional, silencing
this warning entirely.

Original-Fixed: 94310
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664585
Original-Revision: 30cac0394f7b2d173faf20a2de94c1894ad860b7
GitOrigin-RevId: e2e495c8c19cfc935002aa46cbd3d6a63abe183a
Change-Id: I1e498facb05f407a386c31c77d118931e79e5917