[roll] Roll fuchsia [bt][gatt] Add UpdateConnectedPeers method to gatt:GATT
Add a separate "UpdateConnectedPeers" method to gatt::GATT. While this
method could ostensibly be combined into UpdatePeer in a number of ways
(e.g. UpdatePeer could take an optional<PeerId> instead, or take a
vector<PeerId> which could be empty), the semantics around indication
confirmation callbacks are sufficiently different/confusing for this to
be a separate method (ex: for the vector<PeerId> case, is the callback
notified for each indication, or when all of them are complete? If the
former, how will the client know when all indications are complete for
the empty vector without knowledge of all current connections? If the
latter, what's the advantage of that vs. separate methods?).

Original-Fixed: 657
Original-Bug: 36375
Test: Moved arbitrary service registration into GattTest helper method
      for reuse. Added GattTest.{
         NotifyConnectedPeersNoneConnectedDoesntCrash,
         NotifyConnectedPeerWithConnectionDoesntCrash }
      Added GattIndicateMultipleConnectedPeersTest harness as a
      specialization of GattTest for testing indication callback
      semantics.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647003
Original-Revision: 73ec04dbe3f071772c46fbc125fb32c2886c0265
GitOrigin-RevId: 013a0f49e7d4ec50031e35fafc23f193ab287496
Change-Id: I8fe7b751f0614b91116d49f9025892e8a44de129