[roll] Roll fuchsia [stress_tests] Fix broken Component Manager stress test
Component Manager stress tests do not currently run on CQ. A
recent change to the open flags in fuchsia.io broke the compilation
of this test, because it was not updated.

Fixed the broken test.

Added a new CQ variant of this test that should run for a minute
before terminating.

Test: fx test component-manager-stress-tests-alt
      fx test component-manager-stress-tests-short
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664682
Original-Revision: 613e254efde19e93892dd4645affccab895f72fe
GitOrigin-RevId: dae47a460023b4920ad5fca074b9ef7fd8440d33
Change-Id: I27f9c2a707d3e76535234ea9703595098b019ef2