[roll] Roll fuchsia [x86] Fix incorrect return code
Have the board driver return `ZX_OK` when asked to reboot. We
expect the higher layers of the system to handle the reboot by
using Zircon. In the case of a reboot there's no work for the
board driver itself to do.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664885
Original-Revision: be63ba4effc4f75f736b1bf93c63b92409a22525
GitOrigin-RevId: 3552115c4ac5fae0654a688c7fc00ff0b9175194
Change-Id: I235a4889d24e4a7807dd01801bc2b0b5c2aef02e