[roll] Roll fuchsia [fshost] Remove boot arg zircon.system.wait-for-data
If true (which is the default and I cannot find any instances where
it is explicitly set false), zircon.system.wait-for-data guarantees
that fshost's "minfs-not-delayed" directory capability will be serving
the actual contents of minfs before pkgfs is started. If false, there
is no ordering guarantee on when "minfs-not-delayed" changes from being
an empty memfs directory to the actual minfs directory.

In the past, this was useful because some clients (e.g. pkgfs and
many/all components run out of pkgfs) of "minfs-not-delayed"
(historically just called "minfs") assumed that they could access
"minfs-not-delayed" immediately and it would contain the actual minfs
contents.

Now, however, pkgfs no longer uses minfs and all the components that
are run out of pkgfs are given the "minfs" directory capability that
buffers requests until minfs is mounted.

The only remaining clients of "minfs-not-delayed" are console-launcher
and netsvc. These clients launch shells and inject "minfs-not-delayed"
into the shell's namespace as "/data" (so clients of the shell can e.g.
`ls /data`).

There are only two observable behavior changes from this CL:

1. If the boot arg zircon.autorun.system is set, console-launcher will
wait until fshost's "system-delayed" directory capability starts
responding (which means waiting until pkgfs is serving, which means
waiting until minfs is serving, assuming zircon.system.wait-for-data is
true) and then run the supplied command in the shell. Before this
change, setting zircon.system.wait-for-data to true meant that the
zircon.autorun.system shell commands could assume that /data will
contain the actual contents of minfs. There are no known uses of this
behavior.

2. Before this change, if zircon.system.wait-for-data was set to false,
fshost would attempt to launch pkgfs twice, once after blobfs was
launched and again (if the first attempt failed) after minfs was
launched. With this change, fshost will only ever attempt to launch
pkgfs once, after blobfs is launched. This is unlikely to change any
behavior in practice because I think there are no configurations in
which zircon.system.wait-for-data is set to false.

To maintain the property that the delayed directory only starts
servicing requests after blobfs, minfs, and pkgfs have been mounted the
start logic was moved upward from LaunchPkgfs (which is no longer
guaranteed to be called only after minfs is mounted) to
BlockDevice::MountFilesystem.

Original-Fixed: 96533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661891
Original-Revision: a78eb168eb17dc3c83932185e41510909b707959
GitOrigin-RevId: 9e686ccf6cf1dada4f146ef739243b60abe57f43
Change-Id: I654dfc094eddf447b3d5f391bfc5ef9b633e5360