[roll] Roll fuchsia [audio][drivers][aml-g12-tdm] Fix inter-engine dependency
When setting up an audio engine clock configuration it is
possible to override the clock configuration of a different
audio engine since the registers are shared, specifically
when setting the MCLK channel in the EE_AUDIO_MST_PAD_CTRL1
register. This CL fixes the configuration by adding masking
when writing to this register such that configuration for one
engine is not modified when configuring a different engine.

Original-Fixed: b/225428218

Test: fx test -ov aml-g12-tdm-test
Test: Manual audio check using a core build with DAI_TEST
enabled and running multiple audio-driver-ctl instances one
with a tone on the speaker (engine B) and another with a tone
on the PCM used for HFP (engine A). This on Astro.
Test: Manual fixes reproduction of b/225428218, i.e. speaker
audio stops working after ending a Bluetooth HFP call, this on
Sherlock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664850
Original-Revision: c13e0cb250b35b26284b8d01c28dbf100953a2e0
GitOrigin-RevId: e0a7e2a58e410a6fe2a8074fa6564aa2585bff3a
Change-Id: I43be2f4bf2d6deaa4d803db706cef716d71809ed