[roll] Roll fuchsia [cr50_agent] Convert nanos to seconds
A `zx_duration_t` in FIDL is represented as an i64 of nanoseconds.  The
cr50 expects a u32 of whole seconds.  We need to actually convert from
one representation to the other or we'll fail to enroll a credential or
worse, lock users out for a billion times as long as we meant to.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664982
Original-Revision: e544824412a35be5a1e6c538bc02847c664ef20f
GitOrigin-RevId: 475e3f0bcaefb8bc25d7f8df9a1fdf5e94649d8b
Change-Id: I9e82a71cbb3efaa576ef8ee9d2d825160b6f1449