[roll] Roll fuchsia [netstack3] Add receiving dev for packet handling
For SO_BINDTODEVICE, the UDP packet handler needs to be aware of what
device a packet was received on. This CL threads that information
through the stack so that it is available at UDP socket lookup time,
though it is currently unused.

Original-Bug: 96573
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664697
Original-Revision: 9806c6f2c23771ef98d385665d2d3e2b4daee165
GitOrigin-RevId: 8ef90aab7de46f2c49f11c360e45d70a52fc4f6f
Change-Id: I0a5ff937118af43608faefc794aea645de8da284