[roll] Roll fuchsia [starnix][pts] Fix pts node lifecycle
The pts node must be tied to the main terminal being opened and not to
the terminal object being lived. In particular, it must be removed when
the main terminal fd is closed even if some replica are still opened.

The previous code had test passing because the terminal object was not
owned by the replica file descriptors. This is changed in this CL and
the correct lifecycle is ensured.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665065
Original-Revision: 53c1d0bf91c55780d8d5f4520fb482fde399adbd
GitOrigin-RevId: 46067fe512604914dd2177e47e6c4019ad7ee61d
Change-Id: I58289dec82d3aaff08852f63dbc20e2ac95c298c