[roll] Roll fuchsia [test_manager] Decouple ArchiveAccessor in mocks and client logs
Removes the ArchiveAccessor proxy shared between the mock component
and held by test_manager for serving client logs. We are seeing some
odd hangs that are related to logs and seem to be correlated with tests
that access their own diagnostics (and therefore would exercise the
mock). This simplifies the routing so that the mock connects via the
handles it is given, more like a "real" component.
The goal is to simplify the diagnostics system somewhat to either
eliminate or better understand the hangs.

Note: this also moves the mock within the topology somewhat to avoid a
dependency cycle.

Testing: fx test //src/sys/test_manager, test coverage in CI
Original-Bug: 94333

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664066
Original-Revision: e8c7993885aa6617836c19e80975fcebc94423d0
GitOrigin-RevId: f4de13901ee20faa759792a1f4ca73191eef5aaa
Change-Id: I6a410d7d77d2b13d3ca5677cbf49bb696a66596a