[roll] Roll fuchsia [bt][fastpair] Define a Model ID type
The Model ID is assigned to a device during registration. This
varies based on the device and is fixed unless the device is
unregistered in the Nearby Devices Console.

Define a local type for the 24-bit Model ID with conversions to/from
the relevant primitives. This will be used by the GATT service & LE
Advertiser.

Original-Bug: 94166
Test: fx test bt-fastpair-provider-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658789
Original-Revision: 0784c264a79e93e657104697f89a304a8c257ffe
GitOrigin-RevId: 95664e50e9a4d8f765c800869f250d0085af51ac
Change-Id: I71a52630217241696d7bd2130be1d4deb2cabeaa