[roll] Roll fuchsia [fidlc] Warn on attribute typos earlier
This CL moves the attribute typo detection from VerifyAttributesStep to
ConsumeStep. This means these warnings now get reported even if there
are errors during the later compilation steps.

Motivation: most attributes are just extra information that we validate
and pass through to the IR, but there are some that affect compilation,
like @available and @generated_name. These can be load bearing for
correctness: if you remove it, or misspell it (making it a user-defined
attribute), the library fails to compile. An example:

    type Foo = struct {
        // This attribute is crucial to avoid a name collision with Foo.
        foo @generated_name("Bar") struct {};
    };

If you misspell @generated_name, compilation will fail due to the name
collision on "Foo". If you don't spot the typo, this will be confusing,
so this CL ensures we report the typo warning cases like these.

Test: fx test fidl-compiler
 Please enter the commit message for your changes. Lines starting
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664802
Original-Revision: 6064ee5470bffd97ac851ff3e28bcf1acd4a4de5
GitOrigin-RevId: 48b73aba67bdc09e4608c9d0df3cbab1a2d3248c
Change-Id: If282b28a382276ddec12859bfa6acbfb0cc65162