[roll] Roll fuchsia [debugger] Allow async thread controller stops.
This adds the capability to have a thread controller respond that it
wants to respond to the stop later. The same stop will be reissued when
the thread controller reports that it's done.

This is not currently used outside of the newly-added test. This will be
used in a future patch to continue conditionally whether a breakpoint
add has completed or not.

Original-Bug: 95664

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664883
Original-Revision: 48659c27f2b94077ef7e89d637045564699e90ca
GitOrigin-RevId: 7dcbf7918b55f9614cdd7201dc8dbc8ab18f1816
Change-Id: I6b88da11ce5b70b11c260e498cb90d2cb61e4ca5