[roll] Roll fuchsia [netstack3] Perform route lookup without recursion
This is purely a mechanical change. No new feature is introduced and
the stack continues to only allow setting a local route or route through
a gateway without specifying a device (which https://fxbub.dev/96680
will eventually fix).

Note that this change has no performance impact in the lookup/send path
compared to the forwarding table implementation before
https://fuchsia.googlesource.com/fuchsia/+/8a849ee4b9abbfde582be0a7e0e3bd301368991f
which removed the concept of an "active" table. By removing the "active"
table, lookup was no longer O(1) (was O(n) because of the recursive
property on the installed table). This change removes the recursive
property so that lookup operations are O(1) again. Insertion is now
less expensive than the pre-"active" table implementation as we
no longer need to calculate an "active" table on each route table
change.

Original-Fixed: 96162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661726
Original-Revision: 361a340e4a38c548d9ea8a47b2b604f53ccfd470
GitOrigin-RevId: d7f1635f02c73ad2d3ae0cf8962418117b00a965
Change-Id: I8c239bdde9efdf6bb8e101dfe97884cb98d7f620