[roll] Roll fuchsia [debugger] Allow stepping over syscalls.
This implementation detects failures of the "until" controller and falls
back to single-stepping through PLT trampolines. This handles the case
where the destination function is in read-only memory (like the vDSO).

The challenge is that this forces initialization of the until controller
(bubbling up to the Function step thread controller) to be asynchronous.
The previous design didn't allow this.

The extra trickiness is that sometimes initialization is synchronous and
sometimes asynchronous and the return value from OnThreadStop needs to
vary based on this. A new helper class is added to the ThreadController
to handle this complex state combination.

Original-Bug: 95664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664884
Original-Revision: c9ecd9ce106158dda3bf0e58b18898f7d6029f73
GitOrigin-RevId: 0594f390cbb25ba4cbdf83218fb73bc63e983f7a
Change-Id: I1be8e8bed03807cf2587c1515a0d2a3905b49758