[roll] Roll fuchsia [virtualization] Enable SyncWithData test for virtio_block_rs
The device was being overly strict in this situation and sending an
IOERR to the driver when a sync request has data buffers. The spec
indicates that the driver SHOULD NOT do this so we accommodate sync
requests with data buffers.

Original-Bug: 95529
Test: fx test device_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662534
Original-Revision: 09f494661439357d74b9b6b2f69e4a11bc9ac03d
GitOrigin-RevId: e88d02a7a67dc24212904ccfa2d1ce7e74aed528
Change-Id: I84564d783c70445c399092d1d96b34bc68056af5