[roll] Roll fuchsia [netstack3] Replace try_insert() with entry()
Some operations will require examining the current value in the map, and
then inserting if there is none or appending if there is one (and some
other conditions are satisfied). Without entry(), this would have to be
done with a try_insert() then apply(), which duplicates the number of
map lookups being performed.

Original-Bug: 94652

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660662
Original-Revision: 18589c2734dee24663f5e2b8840459274a11ef66
GitOrigin-RevId: 98e255b77ee9a0b24d14b9e4d1e829caa9ca87ef
Change-Id: Iae0a1cbe4fa70418b9684dea5a339881deccb63e