[roll] Roll fuchsia [virtualization] Enable WriteGoodAndBadSectors test for virtio_block_rs.
Update the test to allow the rust device to reject the entire chain if
any part of it is beyond the capacity of the device. We don't change the
behavior of the C++ component since we will likely be removing this
component before too long anyways.

Original-Bug: 95529
Test: fx test device_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663534
Original-Revision: ff84951bc41eebd962e0b59471bcd9d0e623f927
GitOrigin-RevId: c857f74ef5bf59d76c8cb8039c57411988f9e396
Change-Id: Icd24ac8d7fd20467178e7d959cee8a9a39e45f30