[roll] Roll fuchsia [zircon][libc] Use BREAK_ON_LOAD instead of r_brk_on_load
Use ZX_PROP_PROCESS_BREAK_ON_LOAD to store the address of the debug trap
that indicates a change of the dso list. In the past the value is stored
in r_debug.r_brk_on_load, but it's not portable to starnix processes
with other libcs.

This property is not used anywhere yet except in the tests, so existing
behaviors should not change.

Original-Bug: 92960
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652543
Original-Revision: 668fa11cec051e3cd20e8cc14937e24e31c3f552
GitOrigin-RevId: b2643891e341003e765c653b64c7b72eccec031f
Change-Id: I1fc68dba7898b6d878026f11486ae05071db769e