[roll] Roll fuchsia [mdns] add ProxyHostPublisher and ServiceInstancePublisher protocols
This CL adds ProxyHostPublisher and ServiceInstancePublisher protocols
to the mdns service, the latter being a replacement for Publisher.
Quite a bit of new plumbing was required to deal with more specific
transport interaction such as restricting an agent to a particular
medium (wired or wireless) and IP version.

Test: tests added a existing tests updated

Original-Bug: 93994,93320,94918

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656962
Original-Revision: ceb7eaa56ad22c5e95d4fd0afebf483749455731
GitOrigin-RevId: f455cdfe65c53062ba838b4463a5ae3ab11d7464
Change-Id: Ic9c5df7650798edbb0d1e47318a6561a8c04625c