[roll] Roll fuchsia [fshost] Route core the delayed pkgfs directories
Core is stored in pkgfs, so the delayed directories are guaranteed to
be serving before core or anything below it interacts with them.

Also, to highlight the unusual behavior of the not-delayed directories,
add the "-not-delayed" suffix to the not-delayed directories and remove
the "-delayed" suffix from the delayed directories.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663684
Original-Revision: 1a74cb63dec8bb3466bad83002065df4b4625ae3
GitOrigin-RevId: 0c26e9e74c78afa57bfedc52604720fdbd1efd5a
Change-Id: Ibef7bfaf038064c1dd9ed31c1da7094b22a3e435