[roll] Roll fuchsia [fxfs] Track the store object_id of all extent allocations.
This will allow us to safely delete encrypted volumes without needing
the encryption key.

This is not a backwards compatible change because prior to this,
allocation records for different ObjectStore may be merged. After this
change, this is not allowed.

Original-Bug: 89321
Original-Bug: 94221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659483
Original-Revision: 397f7f168590e92236a701fa6d0008c75e27038c
GitOrigin-RevId: 976643502bb6cedd598c79edb7dc1bf4a6f62fb4
Change-Id: I6bcf8967ca0dee79caf38704537743ee0fd1c996