[roll] Roll fuchsia [ffx] ffx component doctor command v0
First iteration for the new `ffx component
doctor` command. Shows outgoing capabilities
that are not exposed and vice-versa.

`exposed` capabilities are declared in the
	manifest.
`outgoing` capabilities are available at
	runtime.

This check is used to verify that the `exposed`
capabilities are available at runtime and that
there are no extra runtime capabilities not
declared in the manifest.

Original-Bug: 94058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654143
Original-Revision: d4f09ac6e38de0af00cca6ae8441b35fe4ddc628
GitOrigin-RevId: c64ceb060077b7b22e006675b01b28c25722a9be
Change-Id: Ibd33cf8dd278aa5ced54eab4bf43cdc6753e918f