[roll] Roll fuchsia [netemul] Wait for test to exit before cleaning up networks
Currently, the netemul runner signals that the test component has
finished running once the test is complete, and then cleans up the
virtual network environment it created for the test. This can cause a
race between network cleanup and realm shutdown: if a component in the
test realm is using one of the virtual endpoints, even though the test
itself is complete, it might log an error or panic if the endpoint is
removed out from under it.

To address this, after closing the component controller channel, the
test runner waits to observe the stopped event from the test root; once
the root is stopped, it can assume that all the constituent components
have been shutdown in an orderly fashion as well. Then it's safe to drop
the network environment.

Multiply: netemul-runner-tests
Multiply: netemul-runner-errors-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666244
Original-Revision: 9a67186db1ae97be91b0661838f240f4f019fc46
GitOrigin-RevId: da84521bd83ff91a23c6f986d534cd1fac0ced7a
Change-Id: I0064f85145150e1cb3f4a466c30ab9e5422c8779