[roll] Roll fuchsia [component][fuzzing] Clean up ComponentContext
This CL modifies ComponentContext to allow constructing a client-only
version that does not consume the directory channel from the startup
handles. When adding the target instrumentation library to a binary to
be fuzzed, it must not consume this handle, or the process will not be
able to serve FIDL protocols of its own.

The CL also refactors the remaining places that were calling
sys::ComponentContext directly with calls to fuzzing::ComponentContext.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662576
Original-Revision: af8070e06c5dc50e9fcc321a0c68c6d80849515a
GitOrigin-RevId: d37ee1046e6fea7e6faca13d6f0bac75988ac978
Change-Id: Ic92d34eef77116dececd6c76c50bbb42c8756fa4