[roll] Roll fuchsia [ftl] Delete read/write fuzzer.
This fuzzer mostly wrote random data to the device and ensured that it
could read the same set of random data back. As a fuzzer it is
essentially just testing the same operation repeatedly with different
data, but that data is not actually processed, it is treated as an
opaque blob. While this not being true would be worrying, it is not a
real issue that is expected to ever come up, and if it does then we've
moved on to doing something very strange.

Original-Fixed: 94146

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666599
Original-Revision: 0d55166e848bb94e7382530969d439e629ddc9d8
GitOrigin-RevId: d7f6ce845f0193bab9a6845155363e2d8d9d7190
Change-Id: I09375bf5338c81ca8a38e7f939ec5cd6a10b616d