[roll] Roll fuchsia [fdf] Add inequality and value lists to device groups
Update the device_group.fidl so that it supports
a property list and conditions for the device groups.

Modify the logic in driver index so that it evaluates
the conditions and value list.

Original-Bug: 96714
Tests: Added unit tests in driver-index-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663862
Original-Revision: b49ea238d0706ff1dc508f254111f72a62499f4f
GitOrigin-RevId: 436db3ff96e67aa1f6a9a974b6fce3d60c8112b1
Change-Id: Id4b4ab1768b4e4147b5aeccc4c14c09d51685a7e