[roll] Roll fuchsia [ffx][pkg] Fall back to use root.json if 1.root.json is not available
Some repositories start with a non-zero root version and expect the client to handle this situation. Unfortunately ffx repository doesn't really support the user to provide this information. As a stopgap, this changes ffx repository to fallback to try the non-versioned `root.json`, which is not formally part of the TUF spec, but most repositories create it, and it corresponds to the latest root metadata.

Original-Bug: 92126
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665943
Original-Revision: 2c423b9a719442321fe1ccc66b4c1979285df09e
GitOrigin-RevId: 7e48016f9babe86bd39ee88988d15066bbaf60ec
Change-Id: Icbda502f3a36ac1e40f1e0eaca6a03288f7005dc