[roll] Roll fuchsia [insntrace] Add "flaky" label for insntrace_tests on NUC
The test is flaky on NUC and blocks the builder. We are disabling the
test on CQ builders until the flakiness issue is solved. The test will
be only run on FYI builders for flaky tests.

Also the test is not supported on FEMU, so we also removed the "emu_env"
from the environments list.

Original-Bug: 97097
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666743
Original-Revision: 51da59f517c98dd1846a8813fb9cf467284d3c51
GitOrigin-RevId: 6809162ab115dd0cb11a3f40a9d1ae2151a502c4
Change-Id: Ib0ebdc46e42ee62eba840358b47e68cc57c4e2d5