[roll] Roll fuchsia [kernel][sched] Fix cluster assignment when favoring LITTLE cores.
Fix an oversight when re-ordering the CPU search sets to favor LITTLE
cores over big cores. The CpuSearchSet::cluster() accessor assumed that
entry 0 is always the CPU the search set is associated with, however,
when ordering the CPUs to favor little cores this is no longer true, as
LITTLE cores should be the first entries for all search sets.

- Update the accessor to use the global cluster map instead of the 0th
  entry in the search set to fix the now incorrect assumption.
- Reduce the cross-cluster migration threshold, which is now being
  correctly applied in heterogeneous systems.

Associating this CL with the power-aware load balancing tracking bug for
context in that effort.

Original-Bug: 96677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666848
Original-Revision: 7c4ffbdbb7f0b5ee208ab114c8326aa596363513
GitOrigin-RevId: e71ea1101ca2886b1f481f00b0453a32635cb6db
Change-Id: Ic8c30d67f009796a1ef72a7f38bebd72c82f35b4