[roll] Roll fuchsia [diagnostics] Make filters match a prefix
Rust ends up generating tags of the form foo::bar, but the filter will
only do an exact match. This CL changes it so that it will do a prefix
match where tags are of the form foo::bar.

This addresses an issue for Fxfs where we can see top-level messages
sent to the console, but no messages below that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665504
Original-Revision: 388092fff25fae0df647da06708e1b01a9f1d475
GitOrigin-RevId: 67a0b28727c47d7c3192843e013fa37ab15094da
Change-Id: Ic069d882b40d16cd0e2094670d958ca14e53652b