[roll] Roll fuchsia [component][fuzzing] Add framework fuzzer-test integration test
This CL adds a fuzzer-test as an integration test. Normally, the GN
templates for component fuzzing (added in a subsequent CL) will generate
both a fuzzer as well as a "fuzzer-test". This test does not use
coverage or mutate inputs. Instead, it verifies the fuzzer can build,
link, and run with a set of fixed inputs. This makes it appropriate for
CQ and helps mitigate bit rot.

This CL explicitly creates a similar "fuzzer-test", but here it is used
to verify the test-engine used with such fuzzer-tests works properly.

In fact, several aspects of building fuzzable target processes have not
been covered by any other tests, and bit-rotted as a result. As a
consequence, this CL also fixes those target library classes as well.

Original-Bug: 92490
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662574
Original-Revision: 6581a25f773a98e44ef99120f21e803175a59774
GitOrigin-RevId: 834f631390f27e19df90a8af98a971b03fb186f4
Change-Id: I1510018a8ceca424f135abc5942f7d1a1b304e60