[roll] Roll fuchsia [kernel] Strip tags in FAR before passing to kernel accesse fault handler
As a followup to fxr/659469, we should also strip the tag when handling
access fault exceptions before they reach the VM's access fault handler.
From what I could tell, there may not be an easy way of toggling the
accessed flag from userspace, but we can manually remove the flag when
aging pages. The `vmaspace_accessed_test` test does this. This CL
reuses some of the machinery for the test such that we can load from
tagged user pointers on arm. The ensures that the FAR is tagged on
access faults.

Original-Bug: 81499
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663583
Original-Revision: 4949feed1eb27c650d2f7b50ad54b8758e03e987
GitOrigin-RevId: 893852e65ca1136e8aeb54b2cd02cb54af9c47cc
Change-Id: I50bae8ebfd6652a6d799ed09dc8358b56690dbae