[roll] Roll fuchsia [hypervisor] Add IdAllocator::Reset
This allows us to dynamically set the size of an IdAllocator to a value
less than or equal to the statically defined maximum.

This is useful in the case where we statically know the maximum possible
value allowed, but must dynamically query the hardware for what limits
it supports.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666435
Original-Revision: 34305c7a218c494be35779deb9f518bf9c443a42
GitOrigin-RevId: b5e488599eb1e1e83558ed60637e1df6d964506c
Change-Id: Icb0d3ad391d435d708c78968abcfb172e3119ccc