[roll] Roll fuchsia [starnix][binder] Restructure locking and waiting
Moves several related fields of data in BinderThread under a single
lock, making it easier to reason about race conditions.

In addition, the waiting mechanism is improved for threads. A WaitQueue
is not necessary, as only a single waiter is ever waiting on a
BinderThread's command queue.

The WaitQueue is needed at the BinderProcess level, so that threads can
wait for events on the /dev/binder FD.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666965
Original-Revision: bd365f583d3be5b2d34b897ce45e4eec8c2b55cf
GitOrigin-RevId: 76787e529fe9dce736f8223fc5c10b8bf2282d19
Change-Id: Ied610b945b575e76a90f9cda2180a66951ff273a