[roll] Roll fuchsia [bt] Add checks and comments to StaticByteBuffer
StaticByteBuffer's ctor is intentionally flexible to allow different
types to be passed to it for silent conversion, so document the
tradeoffs. Also fix the example in its comment.

Add a check for small inputs to verify that no values provided to it are
losing precision through the conversion to uint8_t.

Original-Bug: 41797
Test: fx test //src/connectivity/bluetooth/core/bt-host
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666856
Original-Revision: 9eae1eed11c60d637ce0c33f56f8e491f684ec5b
GitOrigin-RevId: 389744720d1cdbd99efcf8678b9b018deea53aed
Change-Id: I1bdd8d385715992fe43d3ef00d27041961d23667