[roll] Roll fuchsia [pwauth] Drop CredentialManager-managed credentials on account removal
Expand the KeyEnrollment trait to handle both key provisioning and
removal, and then implement removal as no-ops for null auth and
scrypt, and as a call to Credential Manager for pinweaver.

In the event of a failure from Credential Manager, we ignore the failure
and carry on with the account deletion, since our threat model is still
secure against even a malicious Credential Manager provided the user's
password remains secret.

Test: in addition to unit tests, I also performed the following manual
end-to-end test:
* applied the following patch in the series to enable pinweaver and OTAd
* created a new account
* verified via the shell that credential_manager's data directory
  contained a new subfolder under creds/ with three files
* removed the account via `accountctl remove-all`
* observed that credential_manager's data directory no longer contained
  the subfolder of creds/ that had been present before

Original-Fixed: 89078

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666561
Original-Revision: 5ec1ebdd21d9e05491815f3fa2512b8ebf59adc8
GitOrigin-RevId: 1f186d6cb0ecad232b4922e57c211f47c191c759
Change-Id: I7c86d2cee541e6589d2cdb175287545c71821a2f