[roll] Roll fuchsia [starnix] Refactor task interruption and fix exit code race
Task interruption was previously done by abusing the SIGKILl signal.

Instead introduce Task::interrupt that will interrupt a running task and
loop back to the exception loop.
To be noted: This currently only have effect when the task is blocked on
a syscall as the current implementation does not interrupt user-space running code. This
will be changed when it is moved to the new execution model.

This CL also fixes the race between exit and exit_group by ensuring exit
does not override the exit status of a task.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667522
Original-Revision: 3da52bb60ef2d095bf4139cb2a9c3932236b20df
GitOrigin-RevId: ea76247b2d22b5ead8f8a21ac1669211f22c5aee
Change-Id: Ia4c2ff4039480f257ddac9e4226bd41b78dbb81f