[roll] Roll fuchsia [sl4f][cfv2] Delete rotated_square_test.
This test is not being run on any builders despite being included in
src/ui/scenic:test. This is because the rotate_square_test is written as
a host-driven test using SL4F, and it currently builds only as a
dart_library. So I think the builders just builds the test, but won't
run it.

This will allow us to remove the sl4f.scenic.PresentView facade.

Original-Bug: 77056
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667684
Original-Revision: e8b5ad76e5ad2fa60161a9e9409388d893c62088
GitOrigin-RevId: fefc273b15448a0b389bf84e2be84c3104f24ae2
Change-Id: I7ce17a6a380af7bd1f38315c563273cc30fbc707