[roll] Roll fuchsia [products] Introduce hardware_testing product config
This CL removes the hardware_testing bundle and replaces
it with the hardware_testing product config.

This config is presently based on core with several
additional utilities included for the purposes of hardware
testing.

The next step is to set up automated builders. For now
I tested this by building at my workstation, flashing
onto a device and verifying that the necessary utilities
exist in the package cache.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667631
Original-Revision: d4d4cec88fd26fcd818aa7841d15ac64b97bb58d
GitOrigin-RevId: b3d1ae2eab9e03bcf8115ce174b77b32d3faaf22
Change-Id: I9b8ae61f9f0cc368bacb587a1ed153b2399460b0