[roll] Roll fuchsia [insntrace][flake] Ensure device closed before next open
A flake manifested due to the fact that close was not performed
synchronously. The insntrace device only allows for a single client
connection at a time to be open and it was possible for it to observe to
open calls before seeing the close. This race has always been inherently
present, but a recent change must have introduced timing changes which
made it more likely to occur.

MULTIPLY: insntrace_tests
Original-Fixed: 97097
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666955
Original-Revision: 58f834de73f8efa2e942c5b73b25961c4f0f05f3
GitOrigin-RevId: db0367ae77f494c8bb13e95544e3be1eca6eadec
Change-Id: I3468cc680a0e176c203a59a632d0e1914ca39814