[roll] Roll fuchsia [zxdump] Skeleton dump API & gcore command
This implements the skeleton of the dumper logic and the
command-line utility using it.  It just writes empty ELF headers.

The test harness code is extended to sandbox a Fuchsia tool
process so it gets only a private /tmp mapped into a subdirectory
of the test's own /tmp and a private /svc that fakes the
fuchsia.kernel.RootJob protocol with a different job handle.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542803
Original-Revision: 1338ee620cd7780b0365dea01f70035ff6b7c940
GitOrigin-RevId: c85ed8cda2d689697f2758e66f2aba589cfd1adc
Change-Id: Ibfb9156f501b2a5830061dd9f990d4bd3a6f7cb0