[roll] Roll fuchsia [scrutiny] Remove logo from the shell.
This change removes the old logo from the shell. This is a very old
bit of code that was used in some of the initial prototypes. As the
shell becomes involved in more automated processes the logo just
adds extra noise to logs we don't need.

Original-Bug: 83871
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668443
Original-Revision: 21996ba980805bb16a4e7a95c5f10cb199d66f12
GitOrigin-RevId: e01ea0e4cf8824d6ced77e91a56fbdd5df11ada0
Change-Id: I2ac5db3067ae05a2525eaff2813f0099cce5ee22