[roll] Roll fuchsia [CredMgr] Remove or localize allow(dead_code).
Note a chunk of the dead code was for checking the consistency of the
hash tree. I added a consistency check on load to use the functionality
although don't know of any conditions that could cause it to fail.

Also fixed a couple of very small bugs while here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667912
Original-Revision: 311a8d1ff619880be2a4fa01a034bc9c572442e9
GitOrigin-RevId: bb3ce2be7d448d7ad786f7e8cc2eb81054506f43
Change-Id: I146b47ad8c6b0890189e3d4c6d611675d9db1bc0