[roll] Roll fuchsia [CredMgr] Introduce diagnostics trait.
This CL introduces the most basic form of a new diagnostics trait to
record diagnostics events. The trait currently only contains one method
to record the outcome of FIDL methods (i.e. method name plus either
success or the type of error).

Their are two implementations of the trait: a fake that is used in unit
tests (doing this required an additional test to get coverage of the
`handle_requests_for_stream` and `handle_request` methods) and an
inspect implementation that is used in production but is currently
non-functional until the next CL!

Original-Bug: 91714
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668416
Original-Revision: c525af5d23838d1a438f627fb0fdc376e18dc121
GitOrigin-RevId: 4c291a744c3611b0cfaed0eed71161c3f8b76cc5
Change-Id: I005429cdb5df35e864c77ed16c7f8d92488a24e6