[roll] Roll fuchsia [driver_runtime] Add support for delayed tasks
This patch implements support for delayed for tasks by multiplexing a
single delayed task which we post to the shared queue as if it were a
timer object. Unlike a timer, it unforuntately lacks that ability to be
updated, which will cause a slight performance overhead if multiple
concurrent delayed tasks are scheduled on a single dispatcher where the
tasks scheduled later have a sooner deadline than the first. This is in
all likelihood going to be rare as we delayed tasks are not widely used.

I opted to avoid forwarding each outstanding wait with a wrapper to the
shared dispatcher to simplify complexity around cancelation and
shutdown and avoid extraneous allocations.

MULTIPLY: driver_runtime_tests
Test: fx test driver_runtime_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668586
Original-Revision: e5b921d6cee8eb7c1c7fa28d0369a0602acc578a
GitOrigin-RevId: aa32747a2e3af9fb9471dfaea8953bba5bb09b9c
Change-Id: Idf6b68a2d58055cded0507cbf7d41d9c80893ef4