[roll] Roll fuchsia [ddk][fdf] Remove device_get_dispatcher
This was always a temporary measure. Now that we have the driver runtime
available in DFv1 driver hosts, we can use the dispatcher returned by it
instead of manage async loops in the driver host.

As a consequence we no longer spawn a thread per device. These were
largely unused anyways and were a waste of resources.

One outstanding change that is still necessary is a mechanism by which
we can cancel all outstanding tasks for a given device. This behavior is
equivlanet to what we had before, so it's no worse than before.

Original-Fixed: 95885
MULTIPLY: driver_host_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660424
Original-Revision: 5d2b2486dbea56170ed2acba18d49a633544b115
GitOrigin-RevId: 01ad618053e4b661e39ce6f8a5d101340814348d
Change-Id: Iff318dfc4b693bbbd295097cc423e1891b6d5f03