[roll] Roll fuchsia [phys] Trampoline boot relocates data ZBI
When a fixed data load address is passed, the data ZBI is relocated
as well, and the zbi pointer passed to the next kernel item is updated.

This CL is the continuation of the plumbing. Here the Trampoline
code is extended to conditionally copy the data ZBI, when a load
address is provided.

The trampoline test is updated to verify that the boot zbi address
matches the fixed data address.

Also TrampolineArgs is allocated as part of the Trampoline. Since
when doing multiple coies(> 1, due to data and kernel being relocated),
we cannot guarantee that the shim's stack wont be clobbered.

Original-Bug: 88583
Test: fx run-zbi-test trampoline-boot-zbi-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652245
Original-Revision: c35c83c280a2b629d739319c5d093b1f9de7ba8d
GitOrigin-RevId: 3732f23f9bd15bab8ab1690ae6243c1869e858ef
Change-Id: Ia26fc5741c2c84abd813cbb55c2420da2fb25180