[roll] Roll fuchsia [fidl] Transport-dependent iovec array size
Currently, the iovec buffer in UnownedEncodedMessage has a fixed size
regardless of what transport is being used.

This CL adjusts the size based on transport.

This is based on a bug that was discovered with the driver transport
after the FIDL encoder was changed to an alternative encoder.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668404
Original-Revision: 758278fb5ebb9a5920cc3e2143b0ae08ea5868e8
GitOrigin-RevId: eafa9fe9fc9bad8033c7a28727b6a59735055149
Change-Id: I5dff10ea9dd9b4b41e325cafe4d93cb83c2748b9