[roll] Roll fuchsia [driver_runtime] Split Idle event into CompleteShutdown & Idle.
I've changed fdf_internal_wait_until_idle to not wait for waits
and future scheduled tasks.

This is as in the driver host v2 tests, we start up a driver and
want to wait until idle, but it will never succeed if it has called
begin wait but not expected it to be signaled yet.

There are now 2 separate notifications - one for when the dispatcher
enters an idle state (waits_ + timer_ not checked), and the other when
the dispatcher is ready to shutdown (idle state + waits_ empty +
!timer_.is_armed).

Original-Bug: 87162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668856
Original-Revision: 2f5c2da0ff56836c3ae27f475b7afb88ce52c372
GitOrigin-RevId: 55ede6fd9a726f84f8f4a8c9f5fdbad333cc0621
Change-Id: I77dc1955a18d056f3a37cef9963f9d7dce4dea05