[roll] Roll fuchsia [tools][test-list] Add tags from test-list.json to summary.json.
In order to have the test metadata from test-list.json in resultdb, we
can copy it through to the summary.json which we process to upload test
results to resultdb. We can then add those tags to the list of tags we
attach to each test result that we upload.

Original-Bug: b/228092454
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668419
Original-Revision: 5107fb405188a3bc2e94bbe3090e4de3d413d9b0
GitOrigin-RevId: aebabeb77d5e4dad9489bd3e18deb9ec0ec7a1e8
Change-Id: I0a6bff1bc0afaee08df04e9e6841550279517107