[roll] Roll fuchsia [fdf][examples] Write DFv2 demo-number example
This is going to be the example driver for DFv2.
Start by building the driver with the current DFv2
libraries, but this example will be updated to
use best practices and SDK only libraries.

This change was originally reverted because it broke
the build, but the build errors were fixed.

Original-Bug: 97855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669042
Original-Revision: f3ee2ecd447360ec3b074973492856431eda905c
GitOrigin-RevId: a048e05813e16a4d812e7d5df156097ed813f097
Change-Id: I225fb0b61a184b8ae63a1f7eb88dd6e4de88af76