[roll] Roll fuchsia [driver_runtime] Fix DispatcherCoordinator race condition.
When calling the driver notify shutdown handler (usually will be
the driver host), we should not remove the driver_state object
until the handler has completed.

This is as we do not want the shutdown handler to be running,
while another thread thinks the all_drivers_destroyed condition
is true. This could happen if someone calls
fdf_internal_wait_until_all_dispatchers_destroyed or
the last dispatcher of the last driver is destroyed.

Original-Bug: 87162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669028
Original-Revision: 7fba90c4aa844094c36b9353061483a567315915
GitOrigin-RevId: 592a00c873f086bd29bf93b9f151350be5f85f7e
Change-Id: Ib22515c34ccc9dce6e485f8afe61faae257c99f8