[roll] Roll fuchsia [gigaboot] Construct GIC configuration from ACPI
Uses the MADT table to construct either a GICv2 or GICv3 configuration.
We could likely configure a large portion of the x64 interrupt
controller with the same approach, but once again that's autodetected
in the kernel so we don't need to do that here.

Original-Bug: 97287
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667512
Original-Revision: 55eddbd23fd11ae9922b153e84054dff73b7322c
GitOrigin-RevId: 58bec911846cdaf68dcd78e16fe957c35579dc91
Change-Id: Ie5fd088e5dda3c6220efe0dc2c0d7fcdcc527bc7