[roll] Roll fuchsia [weave] Enable weave-ip-fwd-test-wpan-subnet-6
This change updates the prefix length for addresses assigned to the
weave interfaces so that the wlan and wpan nodes are not considered
directly connected to the weave-node. Without this change, weave-node
will not attempt to route packets to the wlan or wpan node through
the fuchsia-node (gateway).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667244
Original-Revision: 178d9d4e6cc184b778cd6219731a837031fe749c
GitOrigin-RevId: 4d0ab681083ca8607c49e41f885cfc1b6aabe0c3
Change-Id: I508e5f0737c3cd9723fcb3be3f0f0af34cf7e047