[roll] Roll fuchsia [tefmocheck] Respect except blocks even if split across tests.
When a failure check has AttributeToTest: true, sometimes the start
string of an except block can show up in one test's output and the end
string shows up in the next test's output. Since we only check the
output for one test at a time, we don't see the end string and so we
count the failure string as a failure when it should be ignored.
Instead, we'll check the entire output log and only consider the failure
strings within the block per test but still check whether it's contained
within an except block in the whole log.

Original-Bug: 97776
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668746
Original-Revision: 00ee8b03d65067c583f7c81cbb624b4992534030
GitOrigin-RevId: 107d55ea679c04a2f9fd7d5edd25339942e173ea
Change-Id: If935925fe9d48845664dea2d0f1c6cfc787d290c