[roll] Roll fuchsia [network/tests] Remove dhcp_interop test
This test requires netemul guest support to be migrated to CFv2, which
has yet to be built out. Given this is the only test that uses the
netemul guest functionality, opt to remove it, with the option to revive
it as a v2 test once the support exists.

Original-Fixed: 97735, 96132
Original-Bug: 87587
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667707
Original-Revision: 975ced70e10bcce61381a82e83ed93e36dc2144c
GitOrigin-RevId: 08b76725d8da0e3ceb79d2b7448a70bc88b02d1e
Change-Id: I31dbec0a4d16dbfd23d803f469e0499c20adfce9