[roll] Roll fuchsia [driver_runtime] Track threads
We centrally track how many threads we have spawned, avoiding spawning
new ones unnecessarily. We ideally want to reset back down to 1 thread
when possible, but that is tricky as it requires joining all threads
when we are running from within one of those threads being joined.

A Reset method was added which can be called from non-driver runtime
managed threads to reset the number of threads back to 1.

Additionally we clamp the overall number of threads we may spawn to 10
in order to reduce resource usage. This is a temporary solution which
needs to be solved with something more robust in the future.

MULTIPLY: driver_runtime_tests
Tested: fx test driver_runtime_tests -- --gtest_repeat=1000 \
           --gtest_break_on_failure
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651170
Original-Revision: 4543fcc4f06cdfdea6614c3f1504b867671d121b
GitOrigin-RevId: 76b2654e36959059fbfd9de1edf990f310b2486d
Change-Id: Ie0b4c0df76619ae6370fa98111f2f810a160ba7e