[roll] Roll fuchsia [basemgr] Rate limit device reboots
This CL modifies basemgr to rate limit reboots triggered
by crashes of the session shell. Prior to this CL, basemgr
would trigger a system reboot if the session shell crashes
4 times within a 1 hour window. This behavior is largely kept
the same except that now the reboot attempts themselves are
rate limited with an exponential backoff. The growth rate
of this limit is by powers of 2, e.g. 2, 4, 8, measured in
minutes since last reboot. The backoff is capped at 64 minutes.
Furthermore, after one day, the backoff resets to 0. Lastly,
logic is added to ensure that this only executes if and when
the system clock has started.

Original-Bug: 82401

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667802
Original-Revision: 5178f1b2951622ab79c4b7c1e3c580ef7534f0f2
GitOrigin-RevId: 451db75b92cc58a3fe3ee43a2fdc6cad0d75ff15
Change-Id: Ic9abdbbc345ccbdc849361212f9ed2b1c060994d