[roll] Roll fuchsia [ffx] Make unknown names unique
We sometimes show multiple targets with the name <unknown>. This adds
some code so that in those cases we show something like <unknown-1> and
<unknown-2> instead. The actually index that we use is the index into
the list of total targets so if it is the first target in the list it
would get <unknown-0> for example.

There is a higher level question about how to make this all better but
this is a good enough fix to help alleviate some pain.

Original-Fixed: 93823

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667002
Original-Revision: 7a0a8087b2ceeb44d108be653a0dbff02fe517be
GitOrigin-RevId: 6d62ff6baeebb9b726f4c2ce679b8fc00e40e33c
Change-Id: Ie2efcc2f3ba1e3f6f3dd01f75dff2162a3d14fbe