[roll] Roll fuchsia [ui][input-pipeline] rename PointerMotionScaleHandler
We're soon going to add another handler which scale pointer
motion, so the name of this handler is going to become
ambiguous.

Accordingly, we rename PointerMotionScaleHandler to
PointerMotionDisplayScaleHandler. This name specifies
the concern that we scale for.

This is purely a rename. There are no functional
changes.

Original-Bug: 88200
Test: fx test input_pipeline_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669676
Original-Revision: 5f6b986164683a3c9679aa95b88af446ab1781cc
GitOrigin-RevId: 253deed7d1af5ed277b1cb2d7a92c052466489fc
Change-Id: I6c1b78f09125a00440e493863382fc4d57d6a7a9