[roll] Roll fuchsia [sysmem] PCMM
Protected Contiguous Memory Management

(Also [aml-securemem])

This is the non-firmware portion of PCMM.  The HW protection ranges
are dynamically managed to un-cover as many pages as feasible and
the un-covered pages are loaned back to Zircon for use in pager-backed
VMOs.

Protected ranges are pinned while protected, to ensure that if sysmem
crashes, Zircon won't be trying to use protected pages.  In a later CL
we'll likely make the system reboot fairly quickly if sysmem crashes.

BUG: 98175
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653018
Original-Revision: 8068732efb74beee3b32756372b4753a02e98db9
GitOrigin-RevId: b76b11e30f45e7f0ff5365eee900b3aae81d7ae8
Change-Id: I7c953e24de1b3922babc96f2b7c157b18a54fcf1