[roll] Roll fuchsia [fidlgen][dart] Support unknown interactions.
Update the Dart bindings to generate support for both sending and
handling unknown interactions on both the server and client side.

These changes to the bindings are only active when fidlc supplies
the new openness/strictness modifiers.

This is a prototype, so the design choices for how the bindings handle
supporting unknown events and unknown method calls are not finalized.
This initial setup demonstrates how the changes could work and provides
a baseline of tests that cover the changes.

Test: fx test fuchsia-pkg://fuchsia.com/dart-bindings-test#meta/dart-bindings-test.cm
Original-Bug: 88366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663743
Original-Revision: 7c2dede85a0cfcfa3629ed9e0af285cf1b930f26
GitOrigin-RevId: 476eac9e0fe1fc7dbd43fd85e5bace36216dbbc3
Change-Id: I5ec60eb27738b6e87aa38ebf934139ce980240fa