[roll] Roll fuchsia [phys] Rename trampoline-boot-shim-test
Since the test and the shim are the same program now,
because the shim does the boot strap step as the test steps,
we renamed the file in this CL.

We keep it separate to preserve file history in gerrit.

No functional changes.

Original-Bug: 88583
Test: fx build && fx run-zbi-test trampoline-boot-zbi-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655182
Original-Revision: 2213c487f2ba45747925acec46ff257b3a5f1efb
GitOrigin-RevId: ddb1cc35170a7fbbaaa47cef1270a2bdf5d6efe8
Change-Id: I3888b4a949a2883e389b9a1c095d92db06d839d9