[roll] Roll fuchsia [ffx] Improve error handling in ffx emu
This CL removes the majority of the "expect" calls from the ffx emu
code, since these cause a panic if the result is an error. This bypasses
much of the surrounding error handling, leading to lost information.

An example of the effects of this change are in fxbug.dev/96781. The
SSH key embedding would fail and the output would just be "no
configuration value found", which comes from a failed call to
ffx_config::file(), but is rather opaque from an end-user's perspective.

This CL also addresses the in-tree/out-of-tree discrepancy when the
product bundle setup fails, as seen in fxbug.dev/95843. We make a check
to the FFX config value sdk.type, and print a different error message
for "in-tree" than not. This should clarify the action steps when an
in'tree image is unsupported by FEMU.

Original-Bug: fxbug.dev/96781
Original-Bug: fxbug.dev/95843
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665108
Original-Revision: 96c29c6babe51ca45925f6a782cced49896e6106
GitOrigin-RevId: 531417d581420ccbe92279f70e8dd052e09d0976
Change-Id: I8f6d915a71ad8e50974978b43e4f0995792a0562