[roll] Roll fuchsia [run-test-suite] Test that streamed artifacts are polled
Adds an integration test that verifies that streamed artifacts
are polled at the same time tests are run. This is important as
artifacts such as logs need to be shown to the developer as they
arrive, rather than once at the end.
To test this, we add a writer that sends asynchronous signals when
certain strings are written. Then we run a hung test, and verify that
any output it produces is visible before cancelling the test.

Testing: new unit and integration test
Multiply: run_test_suite_integration_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669771
Original-Revision: c4bc8490ef8b80334f7ed608745a5baaac9657e6
GitOrigin-RevId: 49ac4fdd3a3769f3e5c7bc9f33fca57662acc027
Change-Id: I290fcd603ec2f3a5d96b7ce4eb7b50009c9358da