[roll] Roll fuchsia [perftest] Change tracing initialisation to be non-racy
Change the perftest library's trace provider startup code to avoid
race conditions in which early trace events can be lost.  That is done
by changing StartTraceProvider() to wait for tracing setup to complete
before returning.

The logic for doing that is somewhat complicated, so in the future it
would be good to move StartTraceProvider() to make it easily usable by
other code.

This logic is loosely based on CreateProviderSynchronouslyAndWait()
and WaitForTracingToStart() in
garnet/bin/trace/tests/integration_test_utils.cc.

Add a test for this which checks for trace events produced by the
perftest library.

Original-Bug: 22911
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/364515
Original-Revision: 343f82b086c3e339ca908494e6d04897257a1e34
GitOrigin-RevId: 1e57b237f33cf7c7615789b91628bb0ea345340b
Change-Id: I746c5d3b4cb8313fab33fa21223b0b5ba0d1a23b