[roll] Roll fuchsia [starnix] Add the ability to run host tests
This CL makes it possible to write tests for starnix in C++ using gtest
in the Fuchsia source tree. These tests are built against the Fuchsia
Linux host toolchain and then run in starmium.

This CL adds a syscalls_test binary, which we can add gtest cases to as
we work on Starnix. The CL also shows how to run the existing fdio-test
host test using this mechanism. For now, the CL enables only a few tests
for this binary. We can enable more of these tests in a later CL.

Test: syscalls_test, fdio-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669947
Original-Revision: 18879b7ae82e6e26a1c992f60a41dd4175a5f2ff
GitOrigin-RevId: c59c4044a3346e7a788ccb94f8a36dcb177e02c2
Change-Id: Id428e34974e6be23af29c0982a8394ab8e292cca