[roll] Roll fuchsia [wlan] Begin to use security feature query in SME.
This change plumbs the security support data into SME, and uses it for
SME's internal logic for determining if WPA3 is supported. A future
change will replace the rest of the uses of the old driver features
enum with the security support query in SME. This test passes the ACTS
test suites PolicyScanTest and VapeInteropTest.

Original-Bug: 88315
Test: fx test wlanstack-tests wlan-hw-sim-test wlan-sme-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669849
Original-Revision: e1406eafaaa3a63785b180df1b89a847531cf22a
GitOrigin-RevId: afd8dfe35300c7539708afea144fbaf547ba802c
Change-Id: I824e423c9b46f92b91e9fda0a253c4a99ff036f5