[roll] Roll fuchsia [fdf] Mark if a matched driver is a fallback
Driver index will include a new boolean field in `MatchedDriverInfo`
called `is_fallback` that will denote whether or not the driver is a
fallback driver. This will be used in a future commit by the driver
loader in order to determine if a matched driver is a fallback driver.

Original-Bug: 92329
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669043
Original-Revision: befff258634e6ee6f59c92fab2b7d9dc582bc3ac
GitOrigin-RevId: ee5b2672891ee315cb9727a6be0fec646e3da6a8
Change-Id: If3eaa020cd2148681459f853bfe917f58aa60344