[roll] Roll fuchsia [kernel][vm] Minorly reorganize AllocateCopyPage() for clarity
The control flow in this function was not structured in a particularly
clear way, and could give off the impression that a variable was
mistakenly used in an uninitialized state. While this was not true, it
serves to restructure the logic a bit for clarity.

Original-Fixed: 97879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669962
Original-Revision: f6a01f36659101169d7e0e1a798099aebc434556
GitOrigin-RevId: b39d3a11da3afbe0b67959741c337d9df4c0b0d2
Change-Id: Iac9802cbf3d6c7c8967f576431ffbec7770c00b6