[roll] Roll fuchsia Revert "[dwc3] Delete unused driver."
This reverts commit bc20f2681a42701797135f417396da0a5aaa1537.

Un-delete an old driver for the DesignWare USB3 core.  Note that this
is currently dead code; it is not hooked into the build system, and
would not build even if you tried.

The next CL will restore the driver to a functioning state, as well as
adding some tests.  The primary purpose of this CL is to provide a
reference as to what changes were made to get the driver working again
after restoring.

Original-Bug: 96652

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667811
Original-Revision: 269ec26767a738b2ce47da6d0f3237e93d1402c8
GitOrigin-RevId: f541c042ff33791ed15ee1137a0438e433db2e68
Change-Id: I70a5f2871a79fedad9552205a3aa51afe3627786