[roll] Roll fuchsia [fx][serve] Fix erroring out with invalid port
In a prior refactor, we accidentally broke `fx serve-updates -l :8084`
from erroring out with the ffx repository server, and the server was not
configured to use that port with `ffx config set
repository.server.listen "[::]:8084"`. This patch changes the helper
function to not unilaterally exit early if the port happens to be free,
but also do the rest of the validation steps.

Original-Fixed: 98017
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670249
Original-Revision: 8a83acd5af855f161f4669e66d82aa5a8f65a532
GitOrigin-RevId: a1aaaa0299387ec9ca2deebfb1900461cf8917b5
Change-Id: I6f913c96162c7516492a1fda64411669c7d8a697