[roll] Roll fuchsia [zbi_test] Set timeout for the host test to 10 min.
Some zbi tests set a timeout of 30s and the success string shows up just
around that timeout, but there's a little delay in seriallistener
reading the success string. Thus, testrunner kills the test just as
seriallistener finds the success string and so it looks like the test
passed but testrunner marks it as timed out.

We will revert the timeout to 10 minutes which was the timeout when the
tests were running on the zbi_tests builders before we ran them as host
tests. Once we have more data, we can determine an appropriate timeout
for the tests.

Original-Bug: 97267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669506
Original-Revision: b6f2f8356021e4e23e235472124015f339911753
GitOrigin-RevId: f237bba75303367f0b7f355e510187aa262cf214
Change-Id: I946dad52f636bbc42547c19de4b0275ef82e4f25