[roll] Roll fuchsia [fake_ddk] Add shutdown of FIDL helper within Ok()
This allows tests to call DdkRelease() after Ok() without the
potential for the FIDL helper to still call code (the FIDL handler)
from the device under test after the device is released.
This surfaced in tests with only one-way calls in which the fake_ddk
would some times call the device under test FIDL handler method after
the device was released since there was no chance for the fake_ddk
thread to call the FIDL handler earlier. With 2-way calls it is always
guaranteed that the FIDL handler is called since the test would block
until the 2-way call is replied to which can only happen after the
FIDL handler is called.
These tests still need to make a 2-way call to make sure the one-way
methods executed, but we should stop serving FIDL from objects that
are expected to be safe for DdkRelease() after Ok().

Original-Bug: 95823

Test: fidl-ddk-test
Test: sa-unittest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660305
Original-Revision: 6720dbbdd8e5a0f19c362419e639f5a0e737bc3f
GitOrigin-RevId: e59238601f17ccefd36f2df05a5a284eebdd0b7d
Change-Id: Idac57a2b4f3d32939d8e861b0074b4b7dfe2f2ff