[roll] Roll fuchsia [phys] TrampolineBoot arm64 support.
Update build rules and ASM to build and test trampoline boot
for arm64 as well.

In arm the boot shim provides a ZBI_TYPE_DEVICETREE, which
is expected, but unnecessary for trampoline boot tests.

Because we are performing memory copies with caches turned off,
and the device tree item size is 1 MB in qemu, this causes huge
performance drops (x86 ~100 iters in 40 secs, while in arm64
we barely get ~20). This is due to a zbi size delta of ~1MB
between arm and x86 (devicetree).

By removing it on the bootstrapping step, we can reduce
the time per iteration, and avoid carrying around an
unnecessary item.

Original-Bug: 88583 #done
Test: fx run-zbi-test trampoline-boot-zbi-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659268
Original-Revision: b3e74248788d914c6003c08f21bd5bcd27846e27
GitOrigin-RevId: 2c6dd08616eeb2a0cfb54c1a15bb7035875c49ae
Change-Id: I0716559102b95fd4ca881857fd8f8e13a6af58cf