[roll] Roll fuchsia [swd-e2e-tests] Pin workstation to a specific build
In fxr/661742, the storage team landed a CL that results in minfs being
rebuilt. This would cause the system to lose the authorized_keys file.
As an attempt to avoid this, this patch pins to a specific build, since
the last green build was from before fxb/661742 landed.

Once we have a green build, this CL can be reverted.

Original-Bug: 95993
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670463
Original-Revision: 1ac6dc6f0e6d462e6ddb11b66b9dae228f6c5268
GitOrigin-RevId: e38e3739b2123ee2afc7b88478555e8e301454f6
Change-Id: I0fc482e6bd0643e735f719af75a34cc6c6b73117