[roll] Roll fuchsia [settings] Add proxy targets and reduce visibility
In preparation for migrating away from //garnet, this CL reduces the
default visibility of all targets in //garnet/bin/setui to only be
visible in the file itself.

Targets that are used externally have proxy targets added that scope
down the visibility to the minimum that's needed.

Original-Fixed: 98204
Test: N/A, just modifying build visibility
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669488
Original-Revision: 39d0e6b88774647769821796ecc9da932dd958d3
GitOrigin-RevId: 40808a3d573038d2bbf2b92450ea1e0775d496d0
Change-Id: I7b8f04456720f1a301f4ccbbc8b3779e21ea5ed1