[roll] Roll fuchsia [guest][tests] Use a nonce for repeated commands
Use a nonce to generate the header and footer expected for repeated
attempts of a single command. In case the earlier attempts eventually
succeed, this prevents the output of an earlier abandoned attempt being
interpreted as input from the current attempt.

Original-Bug: 97722
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669846
Original-Revision: 3e2964ced11eff4289187efb619787e9874ea3be
GitOrigin-RevId: a2fafccaf83871fe370a4bc25eac30eff626a266
Change-Id: Ifa1ec530c26dfeb77d09d0116e1ad918ace97764