[roll] Roll fuchsia [phys] depthcharge-improper-zbi-test
This is able to reproduce the issue in atlas before the mentioned
fix. This is the base, which will be later used for serial setup
verification, that is that the serial item passed to the next kernel
matches the expectations.

Test: fx run-zbi-test depthcharge-improper-zbi-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663563
Original-Revision: 3c093e22dff305dc5b6e4d44657ff3fe694f719a
GitOrigin-RevId: 6b7a103376c3ca6c2ffe9d42041da0e8aa4e354d
Change-Id: I026df2c2f3acec80edafef242fa84ebde6ddd8b6