[roll] Roll fuchsia [kernel][vm] AwaitingClean pages beyond supply_zero_offset
Committed Dirty pages beyond supply_zero_offset are transitioned to
AwaitingClean on WritebackBegin. This ensures that a future write is
correctly intercepted to generate a dirty request.

These pages remain in AwaitingClean for now; they are not transitioned
to Clean on WritebackEnd. That transition will be added only when
supply_zero_offset can be advanced on WritebackEnd (done in a separate
commit).

Test: core-pager-writeback
Original-Bug: 94371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670009
Original-Revision: 8d3082046625f3344ae9a88b80815c6fefb31a74
GitOrigin-RevId: a639cec3440465f593ce9d81839ad2f42f8943b0
Change-Id: Ic69082b00b0c2fd4bd3b908b7e73c7f493f2234c