[roll] Roll fuchsia [fint] Upload build dir manifest to GCS on no-op failure
If the build doesn't converge to a no-op then it's useful to be able to
look at a manifest of the contents of the build directory, including
symlink destinations and mtimes of all files.

The file will contain a JSON array with elements of the form:

  {
    "path": "host_x64/foo",
    "mtime": 1648758742802.102,
    "symlink_target": "../../prebuilt/foo/foo"
  }

"symlink_target" will be omitted if the file is not a symlink.

Original-Bug: 94334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664297
Original-Revision: 61365c1ac79cbce2970b0f64727d4e1b98e74cc0
GitOrigin-RevId: 19d43c3250558b4b3ec5e7bd48ce0becab45d09c
Change-Id: Ife1d897f04357ae0b988e9a7f7c03c152bf699cc