[roll] Roll fuchsia [ffx] Fix sdk-type error for emulator
When adjusting some error reporting code in the emulator startup code,
we overlooked that the in the SDK, a config value is omitted rather than
populated with a different value. This led the launcher to erroneously
fail when launched from the SDK.

This CL removes the invalid check and replaces it with the ffx config
API call which does all the appropriate error handling for us.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670250
Original-Revision: a340028077c7f3b2761b9319de806d16ea2409e2
GitOrigin-RevId: 819ec900c4865adc4c511012f3b144ee1e341a74
Change-Id: I1f67b8bdc90ad64a26723adcfc5223f1b0f8cc58