[roll] Roll fuchsia [ffx] Tweak --log helptext
The helptext for the --log parameter is inaccurate, claiming that by
default all output goes to stdout/stderr. In truth, we set up a log file
in the working directory by default, and print that path to stdout. This
CL adjusts that helptext to indicate this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670583
Original-Revision: 471158a8bf9047132873cf30a2cb739f8fd11169
GitOrigin-RevId: fc75e58f74ef82e3ba42db67d3eb78d17ce05d16
Change-Id: I6834c5c5d1b00aca60560dde0d8a45c2306ecdde