[roll] Roll fuchsia [ffx][repo] Starting server should not check server mode
Back before we had `ffx repository server start`, we'd always start the
server in the background if we had any registered repositories. One of
the confusing errors we could get from that though is that if a user had
`pm serve` running, the repository server would silently fail to start.
The only way to tell this happened was to read the ffx logs.

Since then, we've added explicit controls to starting and stopping the
server, and migrated `fx` and `fserve` over to always use these
commands, rather than assume the ffx repository server is running in the
background. This allows us to report that the server couldn't start, and
help guide the user into how to rectify the issue.

Because of this, we no longer need to have ffx repository block running
the server based off the state of `repository.server.mode`. This patch
removes those checks, which should make it easier for users to try out
ffx repository server.

Original-Fixed: 98349
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669901
Original-Revision: 549430ba08eaea08f4c0cc0ac095a8b4cf3bc62f
GitOrigin-RevId: 44a19a43bb1a3c763586b32687548a7edbcc677a
Change-Id: I864f82327e8f2a5c195b7c95ba17903cf35fc47a