[roll] Roll fuchsia [ffx] Change Result<i32> to Result<()>.
The i32 value in the Ok(_) for this result was unused in main.rs and in
the ffx_plugin_impl functions. The i32 values being set are not actually
read later on to determine the exit code (as can be seen in
https://cs.opensource.google/fuchsia/fuchsia/+/main:src/developer/ffx/lib/errors/src/lib.rs;l=130;drc=6f5b3cc7681a97819f7e58b01df63ef876161a48).

Given these values serve no purpose they have been removed to prevent
confusion to readers.

Original-Bug: 98220

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669835
Original-Revision: 49f9730f89756eb3cd6f90660ecbc33430491d1a
GitOrigin-RevId: 9ec5b4e1802a5a3dc07f21bd0a424d08674965f6
Change-Id: I9360a21a01ab106d21e06c036da998787def2b1c