[roll] Roll fuchsia [media/vnext] allow Threads to be invalid
This change updates the fmlib::Thread class to allow a Thread to be
invalid/uninitialized. This allows a Thread to be used in a consistent
way even when it cannot be initialized immediately or when associated
resources must be released before the Thread goes out of scope.

Test: trivial change
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670740
Original-Revision: 836f8aafb58149aae4b3edc32ed3eb5f43f900fe
GitOrigin-RevId: 4bba6ce30af5047ba0871dfd45197da2d3338a49
Change-Id: I85e84f83d0c4063b3c4bb4a426b5a4c574310478