[roll] Roll llvm-project/libc [libc] Add the implementation of the fflush function.
Note that the underlying flush implementation does not yet fully implement
the POSIX standard. It is complete with respect to the C standard
however. A future change will add the POSIX behavior. It should not affect
the implementation of the fflush function however as the POSIX behavior
will be added in a lower layer.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D124073

GitOrigin-RevId: 71bc6059a9ec6deee2a443cbb3432aa198939e6e
Original-Revision: 915e91145d3fd438447ba7ef43869e0e5560d585
Change-Id: I23ee977f51d59845291b4cb0d67648886bb7f5a6