[roll] Roll fuchsia [fidl][cpp] Sync event handler is exhaustive
When calling `HandleOneEvent`, users expect one of two things to happen:
- Either exactly one of their handler functions is invoked.
- Or `HandleOneEvent` returns an error.

This is not the case today because all the virtual methods in the event
handler class have empty default implementations. This means if the
server sends a known but unhandled event, `HandleOneEvent` will return
OK, but none of the user defined handler will be called. This behavior
is quite brittle, and we'd like to make it safer.

This CL changes methods in the sync event handler interface to pure
virtual, requiring users to acknowledge all possible events, similar to
an exhaustive match in Rust. It surfaced and fixed some latent bugs in
user code.

To retain the ease of adding new events, we support transitional events
without too much ceremony via a default implementation that injects an
error, until FIDL files fully embraces versioning. When a transitional
event is not handled by the user, the error reason is
`fidl::Reason::kUnexpectedMessage`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669834
Original-Revision: d8356afb47a661c2c69e44247e97ca3c38bf63d1
GitOrigin-RevId: 6f8a8167c5b1c6c50ca4ba85d789171c4d516eca
Change-Id: I881c18113be5cdc2d1cb39a440f09376a9978d7f