[roll] Roll fuchsia [graphics][compute][target_archive] Disambiguate target_archive tool name on host
The $(host_toolchain) was conflicting with
`//build/toolchain:linux_64` because the tool names produced by the
target archive script shared the same triple { "cpu", "name", "os" }.

This workaround is to prefix the tool name with "host" when using the
host toolchain.

Testability:

The CL fxrev.dev/658743 revealed this issue and it was failing on CQ.

This CL fixes the issue.

Additionally, inspection of `<out>/tool_paths.json` shows that the
host target archive tool names are now prefixed with "host".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670944
Original-Revision: 9349e42fc181c8437faef8757d67837f51300cfc
GitOrigin-RevId: 47ac1972666979497f51e7b3e46633ede81542f2
Change-Id: I7670b19e8c0dac0c6ce72e737a85aa1b3281da07