[roll] Roll fuchsia [appmgr] Fix incorrect conditional expression
Missing parentheses in assignment-as-condition, resulting in the wrong
operator precedence. I think that this CL is a no-op, and that there is
no functional change.

General style change throughout this long function: scope the `status`
to just the `if` block.

Original-Bug: 98027
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669172
Original-Revision: fa98200740f8e9f4cb71efb96288234aafbd5a18
GitOrigin-RevId: c5f4afcf1a905a5ab1f661206bd4352a56fad207
Change-Id: I778845f61e5de2182e50a90343b32e6284da63a5