[roll] Roll experiences [ermine] Remove dialoginfo after dialog is dismissed.
This change removes a dialoginfo from the list of dialogs after the
corresponding dialog was dismissed. It also fixes an issue when a stuck
app that failed to load (https://fxbug.dev/98083) would leave ermine in
an unresponsive state. This was because the `appIsLaunching` flag was
not reset.

Test: Tested manually
Original-Bug: 98083
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/671143
Original-Revision: 101319a8edeb780cf08c674a37ad5c5be19f07b8
GitOrigin-RevId: 3bd56b7a3414bae79da1573932cfedba647449e1
Change-Id: I861e581c0616a5ddd12cc7f57de27c8511434843