[roll] Roll fuchsia [kernel][socket] Change read off and buf length
This changes how read offsets and lengths are stored in each MBuf.
Previously there was both an offset and length remaining stored in
the MBuf, with the length being the amount of data, from that offset,
remaining. This meant that a read would both increment offset, and
decrement length.

This makes it so that the length is just the length of data in the
MBuf, and there is a single read offset stored in the MBufChain. A
single read offset is sufficient since only one MBuf (the current
buffers_.front()) is actually being read from. To then simplify the
MBuf invariants the length is changed from length remaining from the
offset, to just the length of data in the MBuf.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670883
Original-Revision: 9575d4b3d89a3d739afa5829cac397bc362f92ea
GitOrigin-RevId: a7ceebe3b60c2ee6afeed9b3ac0160a525dfe4ce
Change-Id: Ib4d711715a24223e2bc830a3d863da67aa1bdd30