[roll] Roll fuchsia [python][serialization] Add proper support for default values
When deserializing a dict into an instance of a class, if a field of the
class is not present in the dict, properly detect if it has a default
value, and allow that value to be used instead.

Found while converting a class to use serialization module.

Original-Bug: 98339
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669895
Original-Revision: f937e31c9fcfac8cb6a54632ffba99935c31e51d
GitOrigin-RevId: 8e22c6131c1fbf095ed1c079c0cf51fb692b338b
Change-Id: I0acd571e12faa5326b6c09406ca781e595c0a762