[roll] Roll fuchsia [python][serialization] support use of base classes
Before this, serializing a class with a base-class would only
cause the subclass's fields to be serialized, the base-class's
fields would be omitted.

This was due to typing.get_type_hints() only return the hints
for the class of an instance, not the hints for any of its
base classes.

This adds a DFS through the base classes of a type to build up
the full set of field types.

Original-Bug: 98356
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670447
Original-Revision: 053cc0ae8982288bbe1fb952890bafd5696bde88
GitOrigin-RevId: 877f0a533181e4245f1da8381b7902d0b5c133b8
Change-Id: I7693b583bf5fa92ce28c9bd4f10009f81e1b5e4d