[roll] Roll fuchsia [usb][ums] Disable ums-integration-test
The integration test is flaky, and we're having a hard time resolving
the race. Disabling the test to unwedge the tree while we sort out
what's going on with the virtual bus. UMS doesn't see a lot of delta, so
I'm ok turning off this test for the time being.

Original-Bug: 98566
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671236
Original-Revision: e03dc9dd15996a9c0267ad0006717ed3829ae22d
GitOrigin-RevId: 398efe43125c2164e4a05c7527b08819d8bb65f8
Change-Id: I938963d38114429339fc0109fddf24aad26a0cdb