[roll] Roll fuchsia [kernel][backtrace] Ensure DAP backtrace's frame 0 is accurate
In a backtrace a frame can be either a return address or a precise
location.  Usually, the values are return addresses, but sometimes the
first frame is a precise location.  When building a DAP backtrace the
CPU's PC (a precise location) is added as frame 0.  Prior to this CL,
the symbolizer markup did not indicate whether a frame was a return
address or a precise location so frame 0 of a DAP backtrace was often
misleading or just plain wrong.

Change Backtrace to default to printing frame 0 as a return address.

Add a method to Backtrace so that callers can indicate whether frame 0
is a return address or a precise location.

Use this new method in GetBacktraceFromDapState to ensure that frame 0
of a DAP backtrace is correctly classified as a precise address.

See also this discussion: https://fuchsia-review.googlesource.com/c/fuchsia/+/648093/1/zircon/kernel/lib/lockup_detector/diagnostics.cc#b60

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670980
Original-Revision: 12aaad4c3ba82d1896f995ceaae93e39c1ec1257
GitOrigin-RevId: 02c7481689393d356e8a339c682e7180f321f971
Change-Id: Ib0433a229cbeff6a6e8145ff3269fe602ab7cd27