[roll] Roll fuchsia [fdf][bind] Add instruction enums to DecodeBindRules
This CL updates DecodedBindRules so that it stores
the decoded instructions. It modifies
InstructionVerification so that in addition
to verifying the instruction bytecode, it also
converts it into a set of enums representing the
instruction.

This will make it easier to process the instructions
for matching and debugging.

Original-Bug: 93365
Tests: Updated existing tests with the decoded
       instructions
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671530
Original-Revision: e1fbd0773dc6e15685b4aa06715a96ab0529e1d7
GitOrigin-RevId: 683395a2f293fec49125912bd42b4b6166c22df0
Change-Id: Ie465065202c5f669ee3fb5396c96483cab654ca4