[roll] Roll fuchsia [serve-updates] Fix ffx repo with `fx emu`
When `fx emu` registers itself with fx, it always includes the ssh port,
even if it is 22. Unfortunately because of https://fxbug.dev/98589, ffx
cannot handle target addresses that include the port. This results in
ffx repository being unable to communicate with the target.

To work around this, this patch moves the port filtering to happen
when we call `ffx target repository register`. This allows the fx
scripts to continue to connect to devices running on non-standard ports,
while allowing us to use ffx to interact with the device.

While here, this removes passing `--config ffx_repository=true`, since
all these commands are no longer experimental.

Original-Bug: 98589
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671362
Original-Revision: f872963c6bfdedf8aaa3339ba91eb60b2eceb202
GitOrigin-RevId: 6db79890b86e8d65015ec0dcd73ae7734afaae06
Change-Id: Ia3a6cd2ca3589733f98faa0db5fff453a6692160