[roll] Roll fuchsia [cts][e2e] Move PM tests out of CTS
These tests are not running as compatibility tests, so they do not need
to be in the CTS. This CL moves them back to end_to_end. When pm is
migrated to ffx, those tests should be rewritten as CTS tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672047
Original-Revision: 9022fd6fae8d530e8dddf20dabf5fe12f61385b1
GitOrigin-RevId: df3897ff1f5ec74d7c205d2b9d13e4bd37fedbcb
Change-Id: I1de1601738ae97ac0de05ada0066e42b1dd0959a