[roll] Roll fuchsia [graphics][compute] Update spinel-vk-rs Rust bindings to point at Spinel2
The `spinel-vk-rs` Rust bindings now point to Spinel2.

Additionally, because in-tree Rust currently has no awareness of
Vulkan 1.2, a strawman small C wrapper has been created to hide the
1.2 dependencies.

The Spinel backend has been temporarily removed from Carnelian. This
wrapper will be maintained and improved as changes are landed to
Spinel so that reintegration will be simpler than before.

Note that this is a work-in-progress "helper" library and and when
it's integrated with Carnelian it will be clear which Vulkan state
must be made public.

Testability:

* fx shell run spinel-vk-rs-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658743
Original-Revision: 82d24fddd9e14f1abb17adb30504006181827039
GitOrigin-RevId: c9774ccfe7029713f14f765e6e3ba33f96c96acd
Change-Id: Iacdb497da1920bd74eb7c206082fa032114a0bfd