[roll] Roll fuchsia [test-manager] Close event senders as soon as possible
The latest hang in a coverage builder shows that ffx test is waiting
for more run events from test_manager, but all coverage has already
been drained. test_manager knows that there are no more events to send
to the client when an internal mpsc channel is closed.
This is a speculative fix that closes the channels as early as possible
to minimize the chance of run events hanging.

Original-Bug: 94333

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671786
Original-Revision: dc28014c0a80bf39ddb68cc43f18846fe5bff1f7
GitOrigin-RevId: 90002be8632c645d374ddf1924f472b345a1dbea
Change-Id: Idf7b27ad355f6e0248408802daf98d37190c4578