[roll] Roll fuchsia [benchmarks] Add checks to enforce a naming scheme for performance metrics
Enforce some constraints on the metric names that can be used in the
"test_suite" and "label" fields in fuchsiaperf.json files.

While this could be implemented in catapult_converter (in C++), this
is easier to implement (and easier to write tests for) in Dart.

Change the existing tests for convertResults() to create real
fuchsiaperf files rather than using filenames for files that don't
exist.  This is necessary now that convertResults() reads the files
itself.

Original-Bug: none
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672088
Original-Revision: db71408b270c8a379ba0f283ce1fda75f3fd00b5
GitOrigin-RevId: e5232c5d067e44fcf7eda366e43645c7a5a8cb1c
Change-Id: Ieeeae4d339a7ab8e6138408db91a83c76ecbfc54