[roll] Roll fuchsia [ffx] Add kernel_args flag to ffx emu
The attached bug indicates a high-priority need for the ability to
specify additional arguments when launching an emulator. This was
previously handled via the "--kernel_args" flag in fvdl. We have a
project planned to make configuration easier (see fxbug.dev/98268), but
it won't be complete in time for the training sessions described in the
attached bug.

This CL adds a command line flag to support the old behavior of adding
extra args on the command line that are passed through to the emulator
without additional processing. The flag is marked "experimental" in the
helptext to indicate to end-users that this flag is not intended to be
used in permanent workflows such as infrastructure tests. We intend to
remove this flag when fxbug.dev/98268 is complete.

Original-Bug: fxbug.dev/95278
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672746
Original-Revision: 7705cd88fdf79fa11d9318648bc68dd8c19b7563
GitOrigin-RevId: 7b3a9ec4c6f76efa9b95a08b54a75567f515025f
Change-Id: Ie4977eab4b26ac233d602dbd7e5ad820ddf9cd8b