[roll] Roll fuchsia Reland "[fdf] Set fallback and eager drivers with new way"
This is a reland of commit 1ad2898571a625221a0ffb37efa503641ce3df9b.
Make the driver loader use the `is_fallback` field returned by
MatchDriversV1 when considering if a driver is a fallback driver when
matching drivers to a device.

Originally, the driver loader checked if a driver's version began with
'*' in order to determine if a driver is a fallback driver. Now, the
driver index will check if a driver is a fallback driver instead and
will do so by checking if a driver's component manifest's `fallback`
field is set to true. The driver loader cannot check this. Making
fallback drivers use the new method without first updating the driver
loader to rely on the driver index to identify fallback drivers resulted
in unexpected behavior and is the reason why the original CL was
reverted.

Original change's description:
> [fdf] Set fallback and eager drivers with new way
>
> Set the `fallback` field in all fallback drivers' CML files to true.
> Remove '*' from fallback drivers' version.  Identify eager drivers using
> their component URL and not their name.  This is to avoid incorrect
> ordering of drivers returned by the driver index and processed by the
> driver loader.
>
> Original-Bug: 92329
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654862

Original-Bug: 92329
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667215
Original-Revision: ef92bdad31783ffff707f9d6bed770dc8e33f518
GitOrigin-RevId: 97c8d70376e49e2ddfaa0b7021d25f674df99899
Change-Id: I518e8f674ef708a4ae0745219e86ade07fa3da55