[roll] Roll fuchsia [identity] Verify root hash in the tree is the same as loaded value.
This change calls get_log and verifies the hash in the cr50 is the
same that in the merkle tree. If it isn't the same it fails and fails
out from launching the credential_manager. This is strictly better
than having a sync and having every call fail to the credential
manager.

Original-Bug: 95277
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670008
Original-Revision: 4425bed10d9a52989a7e372ab41fc5df06594abd
GitOrigin-RevId: 170dd6411fbbd6a3152ed52a1c9d5783e78fd9fb
Change-Id: Ie82686e0445360960bdf6976fe00d28fe13f3b2d