[roll] Roll infra/recipes [tree_closer] Don't file a bug if tree already closed
This is a bug I introduced in fxrev.dev/668926, now that we no longer
exit immediately if the tree is already closed, when the closure was
done by the bot.

In that case, we should still check the console status to see if we can
reopen the tree, but should then exit early without filing a bug or
changing the tree status if the console is still unhealthy.

Original-Bug: 97321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/672648
Original-Revision: 9d0b140cb71c2c2a7c56f3a86ba99397399e965a
GitOrigin-RevId: 98f491981a9960d4cffc936be4b5f34250964c26
Change-Id: Idf716c2fdbbb3c3f5863fb9b8e834f066803da01