[roll] Roll fuchsia [starnix][signals] Add support for restarting syscalls
When a signal interrupts a syscall, and the signal action handling the
signal has the flag `SA_RESTART`, the syscall should be restarted if the
syscall supports it.

This change adds the machinery for restarting syscalls, and implements
it for the `read`, `write`, `wait4`, and `waitid` syscalls.

Multiply: test_chromiumos_distro

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671326
Original-Revision: d4938b29e1b0f81f6abc378a7d1b65ed7b061afc
GitOrigin-RevId: 30eea35130e262f7d2b66c5d38c3e3b5d9ec3d87
Change-Id: Ie3fbe14d224de01e5d151e03524613011b31257d