[roll] Roll fuchsia [testrunner] Run snapshot even if we hit a timeout.
The detached bug describes an issue where some tasks time out because
ffx fails to respect the test timeout and hangs. In order to debug, it'd
be helpful to have the snapshot available even if the context was
cancelled due to a timeout.

Original-Bug: 94333
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673090
Original-Revision: 2592e442291056cb244c01d1191ce01304993ece
GitOrigin-RevId: f512c2ffe06ef9428bc3032669e8d02d15c88ebb
Change-Id: Iad802de236631f31fb57ac9847cfd30f2c67ef11