[roll] Roll fuchsia [fidl][docs] Add events handling to the binding spec
Clarify that unhandled events do not constitute as errors.

We don't yet have a spec for the dynamic/runtime behavior aspects of FIDL;
the bindings spec seems to be a reasonable next best alternative.

This was partly motivated by the discussions in
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=77785#c10

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545102
Original-Revision: ddc1f9eceee478daeafe74462d257aa201f71583
GitOrigin-RevId: 86e31c60cafc378cde7c3f35c156ee98d82914fe
Change-Id: If5ed3c3cb0c1341f5540c5f3dc5f8f1ce680cc84