[roll] Roll fuchsia [docs] Add instructions to setup Rust's toolchain
This change adds instructions to the Netstack3 contribution guide for
using the `fuchsia` Rust toolchain with `rustup`. This toolchain is
managed by the Fuchsia team and is the recommended toolchain for
Fuchsia development (as opposed to the `stable` or `nightly`
toolchains).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672565
Original-Revision: 4b77a5307bbc42bd11cdd62622200c50ae65efc5
GitOrigin-RevId: 8edc5f492c53f20a365122f3bdd6f6cca984024b
Change-Id: Ib1101d286aec530662ed94316c1f13fc95f8fc05