[roll] Roll fuchsia [fdf][block] Adding r/w protocol to block
In order to replace the posix Read/Write calls
to Block, we add a function for reading and writing
over the block protocol.  This CL also updates a helper
function in remote_block_device which provides an
interface similar to the pread/pwrite posix calls.

bug: 95835

Testing: the new protocol is tested by way of remote
_block_device in the fvm_integration_test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664888
Original-Revision: 8dfc10038736404fdca053362b74b13d5bb0f770
GitOrigin-RevId: a3fe6bd47caf9dee9a85de9eea521264e1cd9ba7
Change-Id: If31f835a6b684b919448ad3fe1e5f1bee68ce19c