[roll] Roll fuchsia [firmware] add cipd_utils.py script
This script will provide some common functionality to interact with CIPD
prebuilts - up until now we've had a bunch of scripts implement this
functionality separately in one form or another, it'll be better to keep
it in a central place with tests.

To start with, this just provides the "changelog" functionality, which
takes two versions of a CIPD package, along with the source tree, and
generates a changelist of all git repos suitable for use in a git commit
message.

Example usage:

  $ ./cipd_utils.py changelog \
      ~/path/to/repo \
      cipd/package/name \
      old_cipd_version \
      new_cipd_version

  -- Changelist --
  repo1:
  [repo1] commit 1
  [repo1] commit 2

  repo2:
  [repo2] commit 1
  [repo2] commit 2
  [repo2] commit 3

  -- Source Revisions --
  repo1: rev1_B
  repo2: rev2_B

Original-Bug: 99001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672844
Original-Revision: 4208a452873fd380b81225c917d32b29dca3f580
GitOrigin-RevId: baef4c8ce776f6e3a71b5330b43aca34d4d23997
Change-Id: I39a29ba22a040dce6056f148b4502ef4fb91b3f5