[roll] Roll infra/recipes [build_test_upload] Support uploading CIPD packages
Various out-of-tree repos that use the build_test_upload.py recipe want
to start uploading CIPD packages, so we should support that use case.

We want to give the repo as much control over the CIPD package contents
as possible, so we let the repo specify a manifest of files for each
CIPD package it wants to upload, by running a script that prints the
manifest to stdout.

However, individual repos should not be able to upload to arbitrary CIPD
packages. Instead, we want the CIPD packages produced by each builder to
be defined in the infra configs, so that we can do static analysis of
dependency chains between builders.

Unfortunately, the need to declare packages statically precludes us from
using a simpler schema where a single command emits a combined manifests
of the contents of all the CIPD packages to upload (or even using a
single manifest to declare *all* file uploads, including GCS and MOS as
well as CIPD). So instead, we require a separate command to emit the
manifest for each CIPD package.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/673460
Original-Revision: cd55345fd58e24846dec5e086a12414034143b2e
GitOrigin-RevId: 7dffc7ca23a796a393f7e24f9b0ea363df74d243
Change-Id: If3d6664d87dd9d29b02894fa36f4f259f4776564