[roll] Roll fuchsia [starnix] Avoid lock cycle in ProcMapsFile
Take the lock inside the closure in ProcMapsFile::read_at
to avoid a re-entrant lock when writing out the result.
Previously, the lock was taken at the start of read_at,
which meant that it was being held throughout the entirety
of the function, including when the results were written
out to the task's memory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673669
Original-Revision: 8e7bbe022b64db6b667ec3a0c836ecde1f105aa4
GitOrigin-RevId: 1e09927a6a3ddb36c308700d28584574c1268701
Change-Id: I16295800d1747818b024d97f72683db33d0e1204