[roll] Roll fuchsia [assembly] Generate image packages manifest
This change extends the UpdatePackageBuilder to generate the images.json
manifest in the update package that refers to the images stored in the
images packages.

This manifest is currently unused, but this manifest will allow a later
change to remove these image assets from the update package itself.

Original-Fixed: 96745

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672549
Original-Revision: a5f050c02fa885eb5ce28f620675cfe5ab458ae0
GitOrigin-RevId: 79950a5c5f9f03de791e68ce9441b89e9c3a798f
Change-Id: I1de91d3326deec7b5d48953b59e50b07362a22c7