[roll] Roll fuchsia [driver_runtime] Improve safety of dispatcher passed to driver.
Instead of destroying the initial dispatcher passed to the driver
in the shutdown callback, we will destroy it when the driver
object is destructed.

This is as we will be passing an unowned dispatcher to the driver,
and it is better to guarantee the dispatcher will be valid for the
lifetime of the driver, otherwise the driver could attempt to
access an invalid pointer after the dispatcher is shutdown, such
as in its Stop hook.

Original-Bug: 87162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671472
Original-Revision: 132ff5866018a8b6207dbb1adfbfb1394d199865
GitOrigin-RevId: 73964d97399c6050d741189d68ccbacc3b66ef90
Change-Id: I2a7e9d976da8115e92e9354e1c483d63db2b4663