[roll] Roll fuchsia [audio][test] Fix audio_core_hardware_tests flake
These tests wait for audio_core startup, and for it to
identify the default audio input device. The tests then
fail if any device arrival/departure occurs during testing.

However, this could happen while the built-in audio OUTPUT
device was still be in the process of being discovered,
initialized and added. This would not be a problem if the
audio_core service was already started, but in certain
test configurations where this test was the first to
reference audio_core, these tests would often fail.

Now we ignore OnDeviceAdded if the device is an output;
we ignore OnDeviceRemoved if it is not our default audio
capture device; we ignore OnDefaultDeviceChanged if we are
not changing away from our default audio capture device.

This CL also clarifies the warning text that we emit when
we detect that the audio hardware device has a detectable
DC bias (we use more direct language that the hardware
should be retired).

Multiply: audio_core_hardware_tests: 30

Original-Fixed: 81544
Original-Fixed: 90834

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673802
Original-Revision: 5022e917883c50a9d6e446712e1a3c4c62506911
GitOrigin-RevId: adb86d7cc6322d9a61fbf705cc053396100c5257
Change-Id: I897aea4cca1cfea74799cadb7a6542e2e0a9f1a6