[roll] Roll fuchsia [starnix] Refactor directory renames
This moves the check on whether the target is an empty directory when the
source is a directory to FileSystem::rename.

This allows to make the check and the replacement atomic, without taking
a lock on the replaced DirEntry that creates a lock order inversion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673424
Original-Revision: b9c0b05cb7ca001e5a6daecf38f0b671a62cf4a4
GitOrigin-RevId: 924ecd571280ff80191305f7ffc9aad0913df937
Change-Id: I55f214beb0ad3243c135830cf5739c9d938b5e13