[roll] Roll experiences [input] Connects main.dart to an alternate service
We are migrating focus distribution away from Ermine.  Focus
distribution is the responsibility of the input pipeline, which ended up
in Ermine by necessity, the reasons for which no longer exist.

We're now pulling focus distribution to where it's supposed to be,
see:
https://fuchsia-review.googlesource.com/c/fuchsia/+/666844

To do that, we must decouple the two uses of
`fuchsia.ui.focus.FocusChainListenerRegistry` so that we can remove one
of the two, but keep the other.  This change does the decoupling.

Original-Bug: 95901
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/666337
Original-Revision: 10e03b0c090026a8ee7c5a2c1f57316bc8ecae3f
GitOrigin-RevId: ca91120126c691875a076d6276ae95951742290f
Change-Id: I4c99c8b12c702ac71f0a772f6d86c1a957cfddeb