[roll] Roll infra/recipes [git] Copy checkout_from_build_input to separate module
In order to support checking out a Gerrit change and rebasing it on the
correct branch (which is necessary to support release branch CQ), the
`api.git.checkout_from_build_input()` module must make a Gerrit API call
to get the target branch for the change. However, this introduces a
circular dependency between the `gerrit` and `git` recipe modules.

My solution is to split out the git checkout logic into a separate
`git_checkout` module. I also considered putting it in the existing
`checkout` module, but that module is currently Jiri-specific and I
would prefer to keep it that way for simplicity, at least until Jiri
goes away.

This is mostly a copy-paste, with a couple differences. To help view
those differences, I uploaded a temporary patchset (#4) containing an
exact copy-paste of the git api.py functions; you can see the diff here:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/673645/4..6/recipe_modules/git_checkout/api.py

Changes:
1. We now always wrap the checkout steps in a "checkout" nesting for
   consistency with Jiri checkouts, so individual recipes no longer have
   to add nesting. Previously we only did this in
   `checkout_from_build_input()`, and not for checkouts that didn't
   respect the build input.
2. Instead of separate `checkout()` and `checkout_from_build_input()`
   methods, we have a single entrypoint that respects the build input by
   default and only ignores the build input if a `revision` is
   specified, or `ignore_build_input` is set. There are a few times when
   people (myself included) have used the plain `checkout()` method when
   we should have respected the build input, so it makes sense to
   respect the build input by default.

We don't actually respect the change's target branch yet; I'll do that
in a later CL.

Original-Bug: 99125
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/673645
Original-Revision: 621ab4d4b73145431bdea776e55a5f9cecdcd447
GitOrigin-RevId: 65bb1473ef0f5b1dbdb200d0dd980e2fb5c1d667
Change-Id: I2f7bf69eca19e384e4570f015fd4736551efb898